IncompressibleNavierStokes

Documentation for IncompressibleNavierStokes.

IncompressibleNavierStokes.IncompressibleNavierStokes

IncompressibleNavierStokes.AbstractODEMethod

IncompressibleNavierStokes.AbstractODEMethodCache

IncompressibleNavierStokes.AbstractRungeKuttaMethod

IncompressibleNavierStokes.AdamsBashforthCrankNicolsonMethod

IncompressibleNavierStokes.BC

IncompressibleNavierStokes.Force

IncompressibleNavierStokes.Grid

IncompressibleNavierStokes.MomentumCache

IncompressibleNavierStokes.OneLegMethod

IncompressibleNavierStokes.PressureSolver

IncompressibleNavierStokes.TimeStepper

IncompressibleNavierStokes.TimeStepper

IncompressibleNavierStokes.BE11

IncompressibleNavierStokes.CHC3

IncompressibleNavierStokes.CHDIRK3

IncompressibleNavierStokes.CN22

IncompressibleNavierStokes.DOPRI6

IncompressibleNavierStokes.DSRK2

IncompressibleNavierStokes.DSRK3

IncompressibleNavierStokes.DSso2

IncompressibleNavierStokes.HEM3

IncompressibleNavierStokes.HEM5

IncompressibleNavierStokes.ISSPm2

IncompressibleNavierStokes.ISSPs3

IncompressibleNavierStokes.MTE22

IncompressibleNavierStokes.Mid22

IncompressibleNavierStokes.RIA1

IncompressibleNavierStokes.RK33C2

IncompressibleNavierStokes.RK33P2

IncompressibleNavierStokes.RK44

IncompressibleNavierStokes.RK44C2

IncompressibleNavierStokes.RK44C23

IncompressibleNavierStokes.RK44P2

IncompressibleNavierStokes.SDIRK34

IncompressibleNavierStokes.Wray3

IncompressibleNavierStokes.bodyforce

IncompressibleNavierStokes.bodyforce!

IncompressibleNavierStokes.build_operators!

IncompressibleNavierStokes.change_time_stepper

IncompressibleNavierStokes.check_symmetry

IncompressibleNavierStokes.compute_conservation

IncompressibleNavierStokes.convection

IncompressibleNavierStokes.convection!

IncompressibleNavierStokes.convection_components!

IncompressibleNavierStokes.create_boundary_conditions

IncompressibleNavierStokes.create_grid

IncompressibleNavierStokes.create_initial_conditions

IncompressibleNavierStokes.diffusion!

IncompressibleNavierStokes.filter_convection

IncompressibleNavierStokes.filter_convection!

IncompressibleNavierStokes.get_timestep

IncompressibleNavierStokes.get_velocity

IncompressibleNavierStokes.get_vorticity

IncompressibleNavierStokes.initialize!

IncompressibleNavierStokes.initialize_rtp

IncompressibleNavierStokes.interpolate_ν

IncompressibleNavierStokes.is_steady

IncompressibleNavierStokes.ke_viscosity!

IncompressibleNavierStokes.momentum

IncompressibleNavierStokes.momentum!

IncompressibleNavierStokes.momentum_allstage!

IncompressibleNavierStokes.momentum_rom

IncompressibleNavierStokes.nonuniform_grid

IncompressibleNavierStokes.nstage

IncompressibleNavierStokes.ode_method_cache

IncompressibleNavierStokes.operator_averaging!

IncompressibleNavierStokes.operator_convection_diffusion!

IncompressibleNavierStokes.operator_divergence!

IncompressibleNavierStokes.operator_interpolation!

IncompressibleNavierStokes.operator_postprocessing!

IncompressibleNavierStokes.operator_turbulent_diffusion!

IncompressibleNavierStokes.operator_viscosity!

IncompressibleNavierStokes.plot_pressure

IncompressibleNavierStokes.plot_streamfunction

IncompressibleNavierStokes.plot_vorticity

IncompressibleNavierStokes.pressure_additional_solve!

IncompressibleNavierStokes.pressure_poisson!

IncompressibleNavierStokes.rSSPs2

IncompressibleNavierStokes.rSSPs3

IncompressibleNavierStokes.set_bc_vectors!

IncompressibleNavierStokes.solve

IncompressibleNavierStokes.solve

IncompressibleNavierStokes.solve

IncompressibleNavierStokes.solve_steady_ibm!

IncompressibleNavierStokes.solve_steady_ke!

IncompressibleNavierStokes.solve_unsteady_ke!

IncompressibleNavierStokes.solve_unsteady_rom!

IncompressibleNavierStokes.step!

IncompressibleNavierStokes.step!

IncompressibleNavierStokes.step!

IncompressibleNavierStokes.step!

IncompressibleNavierStokes.step!

IncompressibleNavierStokes.strain_tensor

IncompressibleNavierStokes.turbulent_K

IncompressibleNavierStokes.turbulent_viscosity

IncompressibleNavierStokes.update_rtp!

IncompressibleNavierStokes.vorticity!

IncompressibleNavierStokes.IncompressibleNavierStokes — Module

Energy-conserving solvers for the incompressible Navier-Stokes equations.

IncompressibleNavierStokes.AbstractODEMethod — Type

AbstractODEMethod

Abstract ODE method.

IncompressibleNavierStokes.AbstractODEMethodCache — Type

AbstractODEMethodCache

Time stepper cache.

IncompressibleNavierStokes.AbstractRungeKuttaMethod — Type

ExplicitRungeKuttaMethod(A, b, c, r)
ImplicitRungeKuttaMethod(A, b, c, r)

Runge Kutta method.

IncompressibleNavierStokes.AdamsBashforthCrankNicolsonMethod — Type

AdamsBashforthCrankNicolsonMethod(; α₁ = 3 // 2, α₂ = -1 // 2, θ = 1 // 2)

IMEX AB-CN: Adams-Bashforth for explicit convection (parameters α₁ and α₂) and Crank-Nicolson for implicit diffusion (implicitness θ). The method is second order for θ = 1/2.

IncompressibleNavierStokes.BC — Type

BC{T}

Boundary conditions with floating point type T.

IncompressibleNavierStokes.Force — Type

Force{T}

Forcing parameters with floating point type T.

IncompressibleNavierStokes.Grid — Type

Grid{T, N}

Nonuniform cartesian grid of dimension N and floating point type T.

IncompressibleNavierStokes.MomentumCache — Type

MomentumCache

Preallocation structure for terms in the momentum equations.

IncompressibleNavierStokes.OneLegMethod — Type

OneLegMethod(β = 1 // 2)

Explicit one-leg β-method.

IncompressibleNavierStokes.PressureSolver — Type

PressureSolver

Lineare pressure solver for the Poisson equation.

IncompressibleNavierStokes.TimeStepper — Type

TimeStepper

Time stepper for solving ODEs.

IncompressibleNavierStokes.TimeStepper — Method

TimeStepper(method, setup, V₀, p₀) -> TimeStepper

Build associated time stepper from method.

IncompressibleNavierStokes.BE11 — Method

Backward Euler

IncompressibleNavierStokes.CHC3 — Method

Chebyshev quadrature and C(3) satisfied. Note this equals Lobatto IIIA

IncompressibleNavierStokes.CHDIRK3 — Method

Chebyshev based DIRK (not algebraically stable)

IncompressibleNavierStokes.CN22 — Method

Crank-Nicholson

IncompressibleNavierStokes.DOPRI6 — Method

Dormand-Price pair

IncompressibleNavierStokes.DSRK2 — Method

CBM's DSRK2

IncompressibleNavierStokes.DSRK3 — Method

Zennaro's DSRK3

IncompressibleNavierStokes.DSso2 — Method

CBM's DSRKso2

IncompressibleNavierStokes.HEM3 — Method

Brasey and Hairer

IncompressibleNavierStokes.HEM5 — Method

Brasey and Hairer, 5 stage, 4th order

IncompressibleNavierStokes.ISSPm2 — Function

Optimal DIRK SSP schemes of order 2

IncompressibleNavierStokes.ISSPs3 — Function

Optimal DIRK SSP schemes of order 3

IncompressibleNavierStokes.MTE22 — Method

Minimal truncation error 22 method (Heun)

IncompressibleNavierStokes.Mid22 — Method

Midpoint 22 method

IncompressibleNavierStokes.RIA1 — Method

This is implicit Euler

IncompressibleNavierStokes.RK33C2 — Method

RK3 satisfying C(2) for i=3

IncompressibleNavierStokes.RK33P2 — Method

RK3 satisfying the second order condition for the pressure

IncompressibleNavierStokes.RK44 — Method

Classical fourth order

IncompressibleNavierStokes.RK44C2 — Method

RK4 satisfying C(2) for i=3

IncompressibleNavierStokes.RK44C23 — Method

RK4 satisfying C(2) for i=3 and c2=c3

IncompressibleNavierStokes.RK44P2 — Method

RK4 satisfying the second order condition for the pressure (but not third order)

IncompressibleNavierStokes.SDIRK34 — Method

3-stage, 4th order singly diagonally implicit (SSP)

IncompressibleNavierStokes.Wray3 — Method

Wray's RK3

IncompressibleNavierStokes.bodyforce! — Method

bodyforce!(F, ∇F V, t, setup; getJacobian = false)

Compute body force F in momentum equations at velocity points. If getJacobian, also compute ∇F = ∂F/∂V.

IncompressibleNavierStokes.bodyforce — Method

bodyforce(V, t, setup; getJacobian = false)

Body force in momentum equations in Finite Volume setting, so integrated dFx, dFy are the Jacobians ∂Fx/∂V and ∂Fy/∂V.

IncompressibleNavierStokes.build_operators! — Method

build_operators!(setup)

Build discrete operators.

IncompressibleNavierStokes.change_time_stepper — Function

change_time_stepper(stepper, method) -> AbstractTimeStepper

Change ODE method. Return a new stepper.

IncompressibleNavierStokes.check_symmetry — Function

check_symmetry(V, t, setup, ϵ = 1e-14)

Check symmetry of convection operator.

flag = 0: no symmetry error flag = 1: symmetry error

IncompressibleNavierStokes.compute_conservation — Method

compute_conservation(V, t, setup)

Compute mass, momentum and energy conservation properties of velocity field.

IncompressibleNavierStokes.convection! — Method

convection!(c, ∇c, V, ϕ, t, cache, setup, getJacobian = false) -> c, ∇c

Evaluate convective terms c and, optionally, Jacobian ∇c = ∂c/∂V. The convected quantity is ϕ (usually ϕ = V).

IncompressibleNavierStokes.convection — Method

convection(V, ϕ, t, setup, getJacobian = false)

Convenience function for initializing arrays c and ∇c before filling in convection terms.

IncompressibleNavierStokes.convection_components! — Method

convection_components!(c, ∇c, V, ϕ, setup, cache; getJacobian = false, order4 = false)

Compute convection components.

IncompressibleNavierStokes.create_boundary_conditions — Function

create_boundary_conditions(T = Float64; bc_unsteady, bc_type, kwargs...)

Create discrete boundary condtions.

Values should either be scalars or vectors. All values (u, v, p, k, e) are defined at x, y locations, i.e. the corners of pressure volumes, so they cover the entire domain, including corners.

IncompressibleNavierStokes.create_grid — Function

create_grid(Nx, Ny, xlims, ylims, stretch)

Create nonuniform cartesian box mesh mesh xlims × ylims with sizes N and stretch factor.

IncompressibleNavierStokes.create_initial_conditions — Method

V, p = create_initial_conditions(setup)

Create initial vectors.

IncompressibleNavierStokes.diffusion! — Method

diffusion!(d, ∇d, V, t, setup; getJacobian = false)

Evaluate diffusive terms d and optionally Jacobian ∇d = ∂d/∂V.

IncompressibleNavierStokes.filter_convection! — Method

filter_convection!(ū, u, diff_matrix, bc, α)

Construct filter for convective terms.

IncompressibleNavierStokes.filter_convection — Method

filter_convection(u, diff_matrix, bc, α)

Construct filter for convective terms.

IncompressibleNavierStokes.get_timestep — Method

get_timestep(setup)

Estimate time step based on eigenvalues of operators, using Gershgorin.

IncompressibleNavierStokes.get_velocity — Method

get_velocity(V, t, setup)

Get velocity values at pressure points. Interpolate velocities to pressure positions using BMx and BMy (constructed in operator_divergence.jl).

IncompressibleNavierStokes.get_vorticity — Method

get_vorticity(V, t, setup)

Get vorticity from velocity field.

IncompressibleNavierStokes.initialize! — Function

initialize(pressure_solver)

Initialize pressure solver.

IncompressibleNavierStokes.initialize_rtp — Method

initialize_rtp(setup, V, p, t)

Initialize real time plot (RTP).

IncompressibleNavierStokes.interpolate_ν — Method

ν_t_ux, ν_t_uy, ν_t_vx, ν_t_vy = interpolate_ν(ν_t, setup)

interpolate the scalar field νt at pressure locations (xp, yp) to locations needed in computing the diffusive terms, i.e. the ux, uy, vx and v_y locations

IncompressibleNavierStokes.is_steady — Function

is_steady(problem)

Returns true if problem is unsteady.

IncompressibleNavierStokes.ke_viscosity! — Method

ke_viscosity!(setup)

Average (turbulent) viscosity to cell faces

IncompressibleNavierStokes.momentum! — Method

momentum!(F, ∇F, V, ϕ, p, t, setup, cache; getJacobian = false, nopressure = false)

Calculate rhs of momentum equations and, optionally, Jacobian with respect to velocity field.

V: velocity field ϕ: "convection" field: e.g. ∂(ϕx V)/∂x + ∂(ϕy V)/∂y; usually ϕ = V (so ϕx = u, ϕy = v) p: pressure getJacobian: return ∇F = ∂F/∂V nopressure: exclude pressure gradient; in this case input argument p is not used

IncompressibleNavierStokes.momentum — Method

momentum(V, ϕ, p, t, setup; getJacobian = false, nopressure = false)

Convenience function for initializing arrays F and ∇F before filling in momentum terms.

IncompressibleNavierStokes.momentum_allstage! — Method

momentum_allstage!(F, ∇F, V, C, p, t, setup, cache, momentum_cache; getJacobian = false)

Call momentum for multiple (V, p) pairs, as required in implicit RK methods.

IncompressibleNavierStokes.momentum_rom — Method

momentum_rom()

Rhs momentum equation for ROM.

IncompressibleNavierStokes.nonuniform_grid — Function

nonuniform_grid(a, b, N, s = 1)

Create a nonuniform of n points from a to b with a stretch factor of s.

IncompressibleNavierStokes.nstage — Method

nstage(rk_method)

Get number of stages of the Runge-Kutta method.

IncompressibleNavierStokes.ode_method_cache — Function

ode_method_cache(method, args...; kwargs...)

Get time stepper cache for the given ODE method.

IncompressibleNavierStokes.operator_averaging! — Method

Construct averaging operators.

IncompressibleNavierStokes.operator_convection_diffusion! — Method

Construct convection and diffusion operators.

IncompressibleNavierStokes.operator_divergence! — Method

Construct divergence and gradient operator

IncompressibleNavierStokes.operator_interpolation! — Method

Construct averaging operators.

IncompressibleNavierStokes.operator_postprocessing! — Method

operator_postprocessing!(setup)

Construct postprocessing operators such as vorticity.

IncompressibleNavierStokes.operator_turbulent_diffusion! — Method

operator_turbulent_diffusion!(setup)

Average (turbulent) viscosity to cell faces: from ν at xp, yp to ν at ux, uy, vx, vy locations.

See also ke_viscosity.jl.

IncompressibleNavierStokes.operator_viscosity! — Function

operator_viscosity!(model, setup)

Classical turbulence modelling via the diffusive term

IncompressibleNavierStokes.plot_pressure — Method

Plot pressure.

IncompressibleNavierStokes.plot_streamfunction — Method

Plot streamfunction.

IncompressibleNavierStokes.plot_vorticity — Method

plot_vorticity(setup, V, t)

Plot vorticity field.

IncompressibleNavierStokes.pressure_additional_solve! — Method

Additional pressure solve. make the pressure compatible with the velocity field. this should also result in same order pressure as velocity

IncompressibleNavierStokes.pressure_poisson! — Function

pressure_poisson!(solver, p, f, t, setup, tol = 1e-14)

Solve the Poisson equation for the pressure with right hand side f at time t.

We should have sum(f) = 0 for periodic and no-slip BC.

IncompressibleNavierStokes.rSSPs2 — Function

Rational (optimal, low-storage) s-stage 2nd order SSP

IncompressibleNavierStokes.rSSPs3 — Function

Rational (optimal, low-storage) s^2-stage 3rd order SSP

IncompressibleNavierStokes.set_bc_vectors! — Method

set_bc_vectors!(setup, t)

Construct boundary conditions.

IncompressibleNavierStokes.solve — Function

solve(problem, setup, V₀, p₀)

Solve problem with initial state (V₀, p₀).

IncompressibleNavierStokes.solve — Method

solve(steady_state_problem, setup, V₀, p₀)

Solve steady state problem of the Navier-Stokes equations. This saddlepoint system arises from linearization of the convective terms.

IncompressibleNavierStokes.solve — Method

solve(unsteady_problem, setup, V₀, p₀)

Solve unsteady_problem.

IncompressibleNavierStokes.solve_steady_ibm! — Method

solve_steady_ibm!(solution, setup)

Solve ibm.

IncompressibleNavierStokes.solve_steady_ke! — Method

solve_steady_ke!(solution, setup)

Solve ke.

IncompressibleNavierStokes.solve_unsteady_ke! — Method

solve_unsteady_ke!(solution, setup)

Solve ke.

IncompressibleNavierStokes.solve_unsteady_rom! — Method

solve_unsteady_rom!(solution, setup)

Solve ke.

IncompressibleNavierStokes.step! — Function

step!(stepper, Δt)

Take a step of size Δt with the given time stepper.

IncompressibleNavierStokes.step! — Method

step!(stepper::AdamsBashforthCrankNicolsonStepper, Δt)

Perform one time step with Adams-Bashforth for convection and Crank-Nicolson for diffusion.

cₙ₋₁ are the convection terms of tₙ₋₁. Output includes convection terms at tₙ, which will be used in next time step in the Adams-Bashforth part of the method.

Adams-Bashforth for convection and Crank-Nicolson for diffusion formulation:

(u^{n+1} - u^{n})/Δt = -(α₁ c^n + α₂ c^{n-1}) + θ diff^{n+1} + (1-θ) diff^{n} + θ F^{n+1} + (1-θ) F^{n} + θ BC^{n+1} + (1-θ) BC^{n} - G*p + y_p

where BC are boundary conditions of diffusion. This is rewritten as:

(I/Δt - θ D) u^{n+1} = (I/Δt - (1-θ) D) u^{n} - (α₁ c^n + α₂ c^{n-1}) + θ F^{n+1} + (1-θ) F^{n} + θ BC^{n+1} + (1-θ) BC^{n} - G*p + y_p

The LU decomposition of the LHS matrix is precomputed in operator_convection_diffusion.jl.

Note that, in constrast to explicit methods, the pressure from previous time steps has an influence on the accuracy of the velocity.

IncompressibleNavierStokes.step! — Method

step!(stepper::ExplicitRungeKuttaStepper, Δt)

Perform one time step for the general explicit Runge-Kutta method (ERK).

Dirichlet boundary points are not part of solution vector but are prescribed in a strong manner via the u_bc and v_bc functions.

IncompressibleNavierStokes.step! — Method

step!(stepper::ImplicitRungeKuttaStepper, Δt)

Do one time step for implicit Runge-Kutta method.

Unsteady Dirichlet boundary points are not part of solution vector but are prescribed in a "strong" manner via the u_bc and v_bc functions.

IncompressibleNavierStokes.step! — Method

step!(stepper::OneLegStepper, Δt)

Do one time step using one-leg-β-method following symmetry-preserving discretization of turbulent flow. See [Verstappen and Veldman (JCP 2003)] for details, or [Direct numerical simulation of turbulence at lower costs (Journal of Engineering Mathematics 1997)].

Formulation: $\frac{(\beta + 1/2) u^{n+1} - 2 \beta u^{n} + (\beta - 1/2) u^{n-1}}{\Delta t} = F((1 + \beta) u^n - \beta u^{n-1})$

IncompressibleNavierStokes.strain_tensor — Method

strain_tensor(V, t, setup; getJacobian = false, get_S_abs = false)

Evaluate rate of strain tensor S(V) and its magnitude.

IncompressibleNavierStokes.turbulent_K — Function

turbulent_K(model, setup)

Compute the constant part of the turbulent viscosity.

IncompressibleNavierStokes.turbulent_viscosity — Function

turbulent_viscosity(model, setup, S_abs)

Compute turbulent viscosity based on S_abs.

IncompressibleNavierStokes.update_rtp! — Method

update_rtp!(rtp, setup, V, p, t)

Update real time plot.

IncompressibleNavierStokes.vorticity! — Method

vorticity!(ω, V, t, setup)

Compute vorticity values at pressure midpoints. This should be consistent with operator_postprocessing.jl.