API Reference

Grid

IncompressibleNavierStokes.create_grid — Function

create_grid(x, y; order4 = false, T = eltype(x))

Create nonuniform cartesian box mesh xlims × ylims with stretch factors stretch. If order4 is true, a fourth order mesh is created.

create_grid(x, y, z; order4 = false, T = eltype(x))

Create nonuniform cartesian box mesh xlims × ylims × zlims with stretch factors stretch.

Force

IncompressibleNavierStokes.SteadyBodyForce — Type

SteadyBodyForce{T}

Steady (constant) body force.

IncompressibleNavierStokes.UnsteadyBodyForce — Type

UnsteadyBodyForce{T}

Forcing parameters with floating point type T.

Processors

IncompressibleNavierStokes.Logger — Type

Logger(nupdate)

Print time stepping information after every time step.

IncompressibleNavierStokes.RealTimePlotter — Type

RealTimePlotter(; nupdate, fieldname)

Plot the solution every nupdate time steps. Available fieldnames are:

:velocity,

:vorticity,

:streamfunction,

:pressure.

IncompressibleNavierStokes.VTKWriter — Type

VTKWriter(; nupdate, dir = "output", filename = "solution")

Write the solution every nupdate time steps to a VTK file. The resulting Paraview data collection file is stored in "$dir/$filename.pvd".

IncompressibleNavierStokes.QuantityTracer — Type

QuantityTracer(nupdate)

Store scalar quantities (maximum divergence, momentum, kinetic energy) every nupdate time steps.

Problems

IncompressibleNavierStokes.SteadyStateProblem — Type

SteadyStateProblem(setup, V₀, p₀)

Steady state problem. The solution V and p should be such that $\frac{\partial \mathbf{V}}{\partial t} = \mathbf{0}$.

IncompressibleNavierStokes.UnsteadyProblem — Type

UnsteadyProblem(setup, V₀, p₀, (t_start, t_stop))

Unsteady problem with initial conditions V₀, p₀ to be solved from t_start to t_stop.

IncompressibleNavierStokes.is_steady — Function

is_steady(problem)

Returns true if problem is unsteady.

Setup

IncompressibleNavierStokes.Grid — Type

Grid{T,N}()

Nonuniform cartesian grid of dimension N and floating point type T.

IncompressibleNavierStokes.Operators — Type

Operators()

Discrete operators.

IncompressibleNavierStokes.SolverSettings — Type

SolverSettings(;
    pressure_solver = DirectPressureSolver(),
    p_add_solve = true,
    abstol = 1e-10,
    reltol = 1e-14,
    maxiter = 10,
    newton_type = :approximate,
)

Solver settings.

IncompressibleNavierStokes.BC — Type

BC{T}

Boundary conditions with floating point type T.

IncompressibleNavierStokes.Setup — Type

Setup(; viscosity_model, convection_model, grid, force, solver_settings, bc)

Simulation setup.

One-dimensional grids

IncompressibleNavierStokes.stretched_grid — Function

stretched_grid(a, b, N, s = 1)

Create a nonuniform grid of n points from a to b with a stretch factor of s. If s = 1, return a uniform spacing from a to b. Otherwise, return a vector $x \in \mathbb{R}^{N + 1}$ such that $x_n = a + \sum_{i = 1}^n s^{i - 1} h$ for $n = 0, \dots , N$. Setting $x_N = b$ then gives $h = (b - a) \frac{1 - s}{1 - s^N}$, resulting in

\[x_n = a + (b - a) \frac{1 - s^n}{1 - s^N}, \quad n = 0, \dots, N.\]

Note that stretched_grid(a, b, N, s)[n] corresponds to $x_{n - 1}$.

IncompressibleNavierStokes.cosine_grid — Function

cosine_grid(a, b, N)

Create a nonuniform grid of N + 1 points from a to b using a cosine profile, i.e.

\[x_i = a + \frac{1}{2} \left(1 - \cos\left(\pi \frac{i}{n}\right)\right) (b - a), \quad i = 0, \dots, N\]

Pressure solvers

IncompressibleNavierStokes.DirectPressureSolver — Type

DirectPressureSolver()

Direct pressure solver using a LU decomposition.

IncompressibleNavierStokes.CGPressureSolver — Type

CGPressureSolver(abstol, reltol, maxiter)

Conjugate gradients iterative pressure solver.

IncompressibleNavierStokes.FourierPressureSolver — Type

FourierPressureSolver()

Fourier transform pressure solver for periodic domains.

Main driver

IncompressibleNavierStokes.create_boundary_conditions — Function

create_boundary_conditions(u_bc, v_bc[, w_bc]; T = Float64, bc_unsteady, bc_type, kwargs...)

Create discrete boundary condtions.

Values should either be scalars or vectors. All values (u, v, p, k, e) are defined at x, y locations, i.e. the corners of pressure volumes, so they cover the entire domain, including corners.

IncompressibleNavierStokes.build_operators! — Function

build_operators!(setup)

Build discrete operators.

IncompressibleNavierStokes.create_initial_conditions — Function

create_initial_conditions(
    setup,
    t;
    initial_velocity_u,
    initial_velocity_v,
    [initial_velocity_w,]
    initial_pressure = nothing,
)

Create initial vectors at starting time t. If p_initial is a function instead of nothing, calculate compatible IC for the pressure.

IncompressibleNavierStokes.set_bc_vectors! — Function

set_bc_vectors!(setup, t)

Construct boundary conditions.

IncompressibleNavierStokes.solve — Function

solve(problem::SteadyStateProblem; processors = Processor[])

Solve steady state problem of the Navier-Stokes equations. This saddlepoint system arises from linearization of the convective terms.

Each processor is called after every processor.nupdate iteration.

solve(
    problem::UnsteadyProblem, method;
    Δt = nothing,
    n_adapt_Δt = 1,
    processors = Processor[],
    method_startup = nothing,
    nstartup = 1,
)

Solve unsteady problem using method.

The time step is chosen every n_adapt_Δt iteration if Δt is nothing.

For methods that are not self-starting, nstartup startup iterations are performed with method_startup.

Each processor is called after every processor.nupdate time step.

IncompressibleNavierStokes.get_velocity — Function

get_velocity(V, t, setup)

Get velocity values at pressure points. Interpolate velocities to pressure positions using BMx and BMy (and BMz), constructed in operator_divergence.jl.

IncompressibleNavierStokes.momentum! — Function

momentum!(F, ∇F, V, ϕ, p, t, setup, cache; getJacobian = false, nopressure = false)

Calculate rhs of momentum equations and, optionally, Jacobian with respect to velocity field.

V: velocity field

ϕ: convected field: e.g. $\frac{\partial (\phi_x V)}{\partial x} + \frac{\partial (\phi_y V)}{\partial y}$; usually ϕ = V (so ϕx = u, ϕy = v)

p: pressure

getJacobian: return ∇F = ∂F/∂V

nopressure: exclude pressure gradient; in this case input argument p is not used

Plot

IncompressibleNavierStokes.plot_grid — Function

plot_grid(grid)

Plot nonuniform Cartesian grid.

IncompressibleNavierStokes.plot_pressure — Function

plot_pressure(setup, p)

Plot pressure.

IncompressibleNavierStokes.plot_velocity — Function

plot_velocity(setup, p)

Plot velocity.

IncompressibleNavierStokes.plot_vorticity — Function

plot_vorticity(setup, V, t)

Plot vorticity field.

IncompressibleNavierStokes.plot_streamfunction — Function

plot_streamfunction(setup, V, t)

Plot streamfunction.

IncompressibleNavierStokes.plot_tracers — Function

plot_tracers(tracer)

Plot tracer.

Visocosity Models

IncompressibleNavierStokes.LaminarModel — Type

LaminarModel(Re)

Laminar model with Reynolds number Re.

IncompressibleNavierStokes.KEpsilonModel — Type

KEpsilonModel(Re)

K-epsilon model with Reynolds number Re.

IncompressibleNavierStokes.MixingLengthModel — Type

MixingLengthModel(Re)

Mixing-length model with Reynolds number Re and mixing length lm.

IncompressibleNavierStokes.SmagorinskyModel — Type

SmagorinskyModel(Re, C_s = 0.17)

Smagorinsky-Lilly model with Reynolds number Re and constant C_s.

IncompressibleNavierStokes.QRModel — Type

QR(Re)

QR-model with Reynolds number Re.

Convection Models

IncompressibleNavierStokes.NoRegConvectionModel — Type

NoRegConvection()

Unregularized convection model.

IncompressibleNavierStokes.C2ConvectionModel — Type

C2ConvectionModel()

C2 regularization convection model.

IncompressibleNavierStokes.C4ConvectionModel — Type

C4ConvectionModel()

C4 regularization convection model.

IncompressibleNavierStokes.LerayConvectionModel — Type

LerayConvectionModel()

Leray regularization convection model.

ODE methods

IncompressibleNavierStokes.AdamsBashforthCrankNicolsonMethod — Type

AdamsBashforthCrankNicolsonMethod(; α₁ = 3 // 2, α₂ = -1 // 2, θ = 1 // 2)

IMEX AB-CN: Adams-Bashforth for explicit convection (parameters α₁ and α₂) and Crank-Nicolson for implicit diffusion (implicitness θ). The method is second order for θ = 1/2.

IncompressibleNavierStokes.OneLegMethod — Type

OneLegMethod(β = 1 // 2)

Explicit one-leg β-method.

IncompressibleNavierStokes.ExplicitRungeKuttaMethod — Type

ExplicitRungeKuttaMethod(A, b, c, r)

Explicit Runge Kutta method.

IncompressibleNavierStokes.ImplicitRungeKuttaMethod — Type

ImplicitRungeKuttaMethod(A, b, c, r)

Implicit Runge Kutta method.

IncompressibleNavierStokes.runge_kutta_method — Function

runge_kutta_method(A, b, c, r)

Get Runge Kutta method. The function checks whether the method is explicit.

Explicit Methods

Missing docstring.

Missing docstring for FE11. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSP22. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSP42. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSP33. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSP43. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSP104. Check Documenter's build log for details.

IncompressibleNavierStokes.rSSPs2 — Function

Rational (optimal, low-storage) s-stage 2nd order SSP

IncompressibleNavierStokes.rSSPs3 — Function

Rational (optimal, low-storage) s^2-stage 3rd order SSP

IncompressibleNavierStokes.Wray3 — Function

Wray's RK3

Missing docstring.

Missing docstring for RK56. Check Documenter's build log for details.

IncompressibleNavierStokes.DOPRI6 — Function

Dormand-Price pair

Implicit Methods

IncompressibleNavierStokes.BE11 — Function

Backward Euler

IncompressibleNavierStokes.SDIRK34 — Function

3-stage, 4th order singly diagonally implicit (SSP)

IncompressibleNavierStokes.ISSPm2 — Function

Optimal DIRK SSP schemes of order 2

IncompressibleNavierStokes.ISSPs3 — Function

Optimal DIRK SSP schemes of order 3

Half explicit methods

IncompressibleNavierStokes.HEM3 — Function

Brasey and Hairer

Missing docstring.

Missing docstring for HEM3BS. Check Documenter's build log for details.

IncompressibleNavierStokes.HEM5 — Function

Brasey and Hairer, 5 stage, 4th order

Classical Methods

Missing docstring.

Missing docstring for GL1. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GL2. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GL3. Check Documenter's build log for details.

IncompressibleNavierStokes.RIA1 — Function

This is implicit Euler

Missing docstring.

Missing docstring for RIA2. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RIA3. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RIIA1. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RIIA2. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RIIA3. Check Documenter's build log for details.

Missing docstring.

Missing docstring for LIIIA2. Check Documenter's build log for details.

Missing docstring.

Missing docstring for LIIIA3. Check Documenter's build log for details.

Chebyshev methods

IncompressibleNavierStokes.CHDIRK3 — Function

Chebyshev based DIRK (not algebraically stable)

Missing docstring.

Missing docstring for CHCONS3. Check Documenter's build log for details.

IncompressibleNavierStokes.CHC3 — Function

Chebyshev quadrature and C(3) satisfied. Note this equals Lobatto IIIA

Missing docstring.

Missing docstring for CHC5. Check Documenter's build log for details.

Miscellaneous Methods

IncompressibleNavierStokes.Mid22 — Function

Midpoint 22 method

IncompressibleNavierStokes.MTE22 — Function

Minimal truncation error 22 method (Heun)

IncompressibleNavierStokes.CN22 — Function

Crank-Nicholson

Missing docstring.

Missing docstring for Heun33. Check Documenter's build log for details.

IncompressibleNavierStokes.RK33C2 — Function

RK3 satisfying C(2) for i=3

IncompressibleNavierStokes.RK33P2 — Function

RK3 satisfying the second order condition for the pressure

IncompressibleNavierStokes.RK44 — Function

Classical fourth order

IncompressibleNavierStokes.RK44C2 — Function

RK4 satisfying C(2) for i=3

IncompressibleNavierStokes.RK44C23 — Function

RK4 satisfying C(2) for i=3 and c2=c3

IncompressibleNavierStokes.RK44P2 — Function

RK4 satisfying the second order condition for the pressure (but not third order)

DSRK Methods

IncompressibleNavierStokes.DSso2 — Function

CBM's DSRKso2

IncompressibleNavierStokes.DSRK2 — Function

CBM's DSRK2

IncompressibleNavierStokes.DSRK3 — Function

Zennaro's DSRK3

"Non-SSP" Methods of Wong & Spiteri

Missing docstring.

Missing docstring for NSSP21. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NSSP32. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NSSP33. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NSSP53. Check Documenter's build log for details.