API Reference

Force

Missing docstring.

Missing docstring for SteadyBodyForce. Check Documenter's build log for details.

Missing docstring.

Missing docstring for UnsteadyBodyForce. Check Documenter's build log for details.

Processors

Missing docstring.

Missing docstring for Logger. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RealTimePlotter. Check Documenter's build log for details.

Missing docstring.

Missing docstring for VTKWriter. Check Documenter's build log for details.

Missing docstring.

Missing docstring for QuantityTracer. Check Documenter's build log for details.

Problems

Missing docstring.

Missing docstring for SteadyStateProblem. Check Documenter's build log for details.

Missing docstring.

Missing docstring for UnsteadyProblem. Check Documenter's build log for details.

Missing docstring.

Missing docstring for is_steady. Check Documenter's build log for details.

Setup

Missing docstring.

Missing docstring for Grid. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Operators. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SolverSettings. Check Documenter's build log for details.

Missing docstring.

Missing docstring for BoundaryConditions. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Setup. Check Documenter's build log for details.

One-dimensional grids

IncompressibleNavierStokes.stretched_grid — Function

stretched_grid(a, b, N, s = 1)

Create a nonuniform grid of n points from a to b with a stretch factor of s. If s = 1, return a uniform spacing from a to b. Otherwise, return a vector $x \in \mathbb{R}^{N + 1}$ such that $x_n = a + \sum_{i = 1}^n s^{i - 1} h$ for $n = 0, \dots , N$. Setting $x_N = b$ then gives $h = (b - a) \frac{1 - s}{1 - s^N}$, resulting in

\[x_n = a + (b - a) \frac{1 - s^n}{1 - s^N}, \quad n = 0, \dots, N.\]

Note that stretched_grid(a, b, N, s)[n] corresponds to $x_{n - 1}$.

IncompressibleNavierStokes.cosine_grid — Function

cosine_grid(a, b, N)

Create a nonuniform grid of N + 1 points from a to b using a cosine profile, i.e.

\[x_i = a + \frac{1}{2} \left(1 - \cos\left(\pi \frac{i}{n}\right)\right) (b - a), \quad i = 0, \dots, N\]

Pressure solvers

Missing docstring.

Missing docstring for DirectPressureSolver. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CGPressureSolver. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FourierPressureSolver. Check Documenter's build log for details.

Main driver

Missing docstring.

Missing docstring for create_initial_conditions. Check Documenter's build log for details.

Missing docstring.

Missing docstring for set_bc_vectors!. Check Documenter's build log for details.

IncompressibleNavierStokes.solve — Function

solve(
    problem::SteadyStateProblem;
    jacobian_type = :newton,
    npicard = 2,
    abstol = 1e-10,
    maxiter = 10,
)

Solve steady state problem of the Navier-Stokes equations. This saddlepoint system arises from linearization of the convective terms.

Each processor is called after every processor.nupdate iteration.

solve(
    problem::UnsteadyProblem,
    method;
    pressure_solver = DirectPressureSolver(problem.setup),
    Δt = nothing,
    cfl = 1,
    n_adapt_Δt = 1,
    nstartup = 1,
    method_startup = nothing,
    inplace = false,
    processors = Processor[],
)

Solve unsteady problem using method.

The time step is chosen every n_adapt_Δt iteration with CFL-number cfl if Δt is nothing.

For methods that are not self-starting, nstartup startup iterations are performed with method_startup.

Each processor is called after every processor.nupdate time step.

Missing docstring.

Missing docstring for get_velocity. Check Documenter's build log for details.

IncompressibleNavierStokes.momentum! — Function

momentum!(F, ∇F, V, ϕ, p, t, setup, cache; getJacobian = false, nopressure = false)

Calculate rhs of momentum equations and, optionally, Jacobian with respect to velocity field.

V: velocity field

ϕ: convected field: e.g. $\frac{\partial (\phi_x V)}{\partial x} + \frac{\partial (\phi_y V)}{\partial y}$; usually ϕ = V (so ϕx = u, ϕy = v)

p: pressure

getJacobian: return ∇F = ∂F/∂V

nopressure: exclude pressure gradient; in this case input argument p is not used

newton_factor

Mutating/non-allocating/in-place version.

See also momentum.

Plot

Missing docstring.

Missing docstring for plot_grid. Check Documenter's build log for details.

Missing docstring.

Missing docstring for plot_pressure. Check Documenter's build log for details.

Missing docstring.

Missing docstring for plot_velocity. Check Documenter's build log for details.

Missing docstring.

Missing docstring for plot_vorticity. Check Documenter's build log for details.

Missing docstring.

Missing docstring for plot_streamfunction. Check Documenter's build log for details.

IncompressibleNavierStokes.plot_tracers — Function

plot_tracers(tracer)

Plot tracer.

Visocosity Models

Missing docstring.

Missing docstring for LaminarModel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for KEpsilonModel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MixingLengthModel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SmagorinskyModel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for QRModel. Check Documenter's build log for details.

Convection Models

Missing docstring.

Missing docstring for NoRegConvectionModel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for C2ConvectionModel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for C4ConvectionModel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for LerayConvectionModel. Check Documenter's build log for details.