API Reference – Runge-Kutta methods

IncompressibleNavierStokes.RKMethods — Module

RKMethods

Set up Butcher arrays A, b, and c, as well as and SSP coefficient r. For families of methods, optional input s is the number of stages.

Original (MATLAB) by David Ketcheson, extended by Benjamin Sanderse.

Explicit Methods

IncompressibleNavierStokes.RKMethods.FE11 — Function

FE11(; kwargs...)

FE11.

IncompressibleNavierStokes.RKMethods.SSP22 — Function

SSP22(; kwargs...)

SSP22.

IncompressibleNavierStokes.RKMethods.SSP42 — Function

SSP42(; kwargs...)

SSP42.

IncompressibleNavierStokes.RKMethods.SSP33 — Function

SSP33(; kwargs...)

SSP33.

IncompressibleNavierStokes.RKMethods.SSP43 — Function

SSP43(; kwargs...)

SSP43.

IncompressibleNavierStokes.RKMethods.SSP104 — Function

SSP104(; kwargs...)

SSP104.

IncompressibleNavierStokes.RKMethods.rSSPs2 — Function

rSSPs2(s = 2; kwargs...)

Rational (optimal, low-storage) s-stage 2nd order SSP.

IncompressibleNavierStokes.RKMethods.rSSPs3 — Function

rSSPs3(s = 4; kwargs...)

Rational (optimal, low-storage) s^2-stage 3rd order SSP.

IncompressibleNavierStokes.RKMethods.Wray3 — Function

Wray3(; kwargs...)

Wray's RK3.

IncompressibleNavierStokes.RKMethods.RK56 — Function

RK56(; kwargs...)

RK56.

IncompressibleNavierStokes.RKMethods.DOPRI6 — Function

DOPRI6(; kwargs...)

Dormand-Price pair.

Implicit Methods

IncompressibleNavierStokes.RKMethods.BE11 — Function

BE11(; kwargs...)

Backward Euler.

IncompressibleNavierStokes.RKMethods.SDIRK34 — Function

SDIRK34(; kwargs...)

3-stage, 4th order singly diagonally implicit (SSP).

IncompressibleNavierStokes.RKMethods.ISSPm2 — Function

ISSPm2(s = 1; kwargs...)

Optimal DIRK SSP schemes of order 2.

IncompressibleNavierStokes.RKMethods.ISSPs3 — Function

ISSPs3(s = 2; kwargs...)

Optimal DIRK SSP schemes of order 3.

Half explicit methods

IncompressibleNavierStokes.RKMethods.HEM3 — Function

HEM3(; kwargs...)

Brasey and Hairer.

IncompressibleNavierStokes.RKMethods.HEM3BS — Function

HEM3BS(; kwargs...)

HEM3BS.

IncompressibleNavierStokes.RKMethods.HEM5 — Function

HEM5(; kwargs...)

Brasey and Hairer, 5 stage, 4th order.

Classical Methods

IncompressibleNavierStokes.RKMethods.GL1 — Function

GL1(; kwargs...)

GL1.

IncompressibleNavierStokes.RKMethods.GL2 — Function

GL2(; kwargs...)

GL2.

IncompressibleNavierStokes.RKMethods.GL3 — Function

GL3(; kwargs...)

GL3.

IncompressibleNavierStokes.RKMethods.RIA1 — Function

RIA1(; kwargs...)

This is implicit Euler.

IncompressibleNavierStokes.RKMethods.RIA2 — Function

RIA2(; kwargs...)

RIA2.

IncompressibleNavierStokes.RKMethods.RIA3 — Function

RIA3(; kwargs...)

RIA3.

IncompressibleNavierStokes.RKMethods.RIIA1 — Function

RIIA1(; kwargs...)

RIIA1.

IncompressibleNavierStokes.RKMethods.RIIA2 — Function

RIIA2(; kwargs...)

RIIA2.

IncompressibleNavierStokes.RKMethods.RIIA3 — Function

RIIA3(; kwargs...)

RIIA3.

IncompressibleNavierStokes.RKMethods.LIIIA2 — Function

LIIIA2(; kwargs...)

LIIIA2.

IncompressibleNavierStokes.RKMethods.LIIIA3 — Function

LIIIA3(; kwargs...)

LIIIA3.

Chebyshev methods

IncompressibleNavierStokes.RKMethods.CHDIRK3 — Function

CHDIRK3(; kwargs...)

Chebyshev based DIRK (not algebraically stable).

IncompressibleNavierStokes.RKMethods.CHCONS3 — Function

CHCONS3(; kwargs...)

CHCONS3.

IncompressibleNavierStokes.RKMethods.CHC3 — Function

CHC3(; kwargs...)

Chebyshev quadrature and C(3) satisfied. Note this equals Lobatto IIIA.

IncompressibleNavierStokes.RKMethods.CHC5 — Function

CHC5(; kwargs...)

CHC5.

Miscellaneous Methods

IncompressibleNavierStokes.RKMethods.Mid22 — Function

Mid22(; kwargs...)

Midpoint 22 method.

IncompressibleNavierStokes.RKMethods.MTE22 — Function

MTE22(; kwargs...)

Minimal truncation error 22 method (Heun).

IncompressibleNavierStokes.RKMethods.CN22 — Function

CN22(; kwargs...)

Crank-Nicholson.

IncompressibleNavierStokes.RKMethods.Heun33 — Function

Heun33(; kwargs...)

Heun33.

IncompressibleNavierStokes.RKMethods.RK33C2 — Function

RK33C2(; kwargs...)

RK3 satisfying C(2) for i=3.

IncompressibleNavierStokes.RKMethods.RK33P2 — Function

RK33P2(; kwargs...)

RK3 satisfying the second order condition for the pressure.

IncompressibleNavierStokes.RKMethods.RK44 — Function

RK44(; kwargs...)

Classical fourth order.

IncompressibleNavierStokes.RKMethods.RK44C2 — Function

RK44C2(; kwargs...)

RK4 satisfying C(2) for i=3.

IncompressibleNavierStokes.RKMethods.RK44C23 — Function

RK44C23(; kwargs...)

RK4 satisfying C(2) for i=3 and c2=c3.

IncompressibleNavierStokes.RKMethods.RK44P2 — Function

RK44P2(; kwargs...)

RK4 satisfying the second order condition for the pressure (but not third order).

DSRK Methods

IncompressibleNavierStokes.RKMethods.DSso2 — Function

DSso2(; kwargs...)

CBM's DSRKso2.

IncompressibleNavierStokes.RKMethods.DSRK2 — Function

DSRK2(; kwargs...)

CBM's DSRK2.

IncompressibleNavierStokes.RKMethods.DSRK3 — Function

DSRK3(; kwargs...)

Zennaro's DSRK3.

"Non-SSP" Methods of Wong & Spiteri

IncompressibleNavierStokes.RKMethods.NSSP21 — Function

NSSP21(; kwargs...)

NSSP21.

IncompressibleNavierStokes.RKMethods.NSSP32 — Function

NSSP32(; kwargs...)

NSSP32.

IncompressibleNavierStokes.RKMethods.NSSP33 — Function

NSSP33(; kwargs...)

NSSP33.

IncompressibleNavierStokes.RKMethods.NSSP53 — Function

NSSP53(; kwargs...)

NSSP53.