API Reference – time steppers

ODE methods

Missing docstring.

Missing docstring for AdamsBashforthCrankNicolsonMethod. Check Documenter's build log for details.

Missing docstring.

Missing docstring for OneLegMethod. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ExplicitRungeKuttaMethod. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ImplicitRungeKuttaMethod. Check Documenter's build log for details.

IncompressibleNavierStokes.runge_kutta_method — Function

runge_kutta_method(A, b, c, r; [p_add_solve], [newton_type], [maxiter], [abstol], [reltol])

Get Runge Kutta method. The function checks whether the method is explicit.

p_add_solve: whether to add a pressure solve step to the method.

For implicit RK methods: newton_type, maxiter, abstol, reltol.

Explicit Methods

Missing docstring.

Missing docstring for FE11. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSP22. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSP42. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSP33. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSP43. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSP104. Check Documenter's build log for details.

IncompressibleNavierStokes.rSSPs2 — Function

Rational (optimal, low-storage) s-stage 2nd order SSP

IncompressibleNavierStokes.rSSPs3 — Function

Rational (optimal, low-storage) s^2-stage 3rd order SSP

IncompressibleNavierStokes.Wray3 — Function

Wray's RK3

Missing docstring.

Missing docstring for RK56. Check Documenter's build log for details.

IncompressibleNavierStokes.DOPRI6 — Function

Dormand-Price pair

Implicit Methods

IncompressibleNavierStokes.BE11 — Function

Backward Euler

IncompressibleNavierStokes.SDIRK34 — Function

3-stage, 4th order singly diagonally implicit (SSP)

IncompressibleNavierStokes.ISSPm2 — Function

Optimal DIRK SSP schemes of order 2

IncompressibleNavierStokes.ISSPs3 — Function

Optimal DIRK SSP schemes of order 3

Half explicit methods

IncompressibleNavierStokes.HEM3 — Function

Brasey and Hairer

Missing docstring.

Missing docstring for HEM3BS. Check Documenter's build log for details.

IncompressibleNavierStokes.HEM5 — Function

Brasey and Hairer, 5 stage, 4th order

Classical Methods

Missing docstring.

Missing docstring for GL1. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GL2. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GL3. Check Documenter's build log for details.

IncompressibleNavierStokes.RIA1 — Function

This is implicit Euler

Missing docstring.

Missing docstring for RIA2. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RIA3. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RIIA1. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RIIA2. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RIIA3. Check Documenter's build log for details.

Missing docstring.

Missing docstring for LIIIA2. Check Documenter's build log for details.

Missing docstring.

Missing docstring for LIIIA3. Check Documenter's build log for details.

Chebyshev methods

IncompressibleNavierStokes.CHDIRK3 — Function

Chebyshev based DIRK (not algebraically stable)

Missing docstring.

Missing docstring for CHCONS3. Check Documenter's build log for details.

IncompressibleNavierStokes.CHC3 — Function

Chebyshev quadrature and C(3) satisfied. Note this equals Lobatto IIIA

Missing docstring.

Missing docstring for CHC5. Check Documenter's build log for details.

Miscellaneous Methods

IncompressibleNavierStokes.Mid22 — Function

Midpoint 22 method

IncompressibleNavierStokes.MTE22 — Function

Minimal truncation error 22 method (Heun)

IncompressibleNavierStokes.CN22 — Function

Crank-Nicholson

Missing docstring.

Missing docstring for Heun33. Check Documenter's build log for details.

IncompressibleNavierStokes.RK33C2 — Function

RK3 satisfying C(2) for i=3

IncompressibleNavierStokes.RK33P2 — Function

RK3 satisfying the second order condition for the pressure

IncompressibleNavierStokes.RK44 — Function

Classical fourth order

IncompressibleNavierStokes.RK44C2 — Function

RK4 satisfying C(2) for i=3

IncompressibleNavierStokes.RK44C23 — Function

RK4 satisfying C(2) for i=3 and c2=c3

IncompressibleNavierStokes.RK44P2 — Function

RK4 satisfying the second order condition for the pressure (but not third order)

DSRK Methods

IncompressibleNavierStokes.DSso2 — Function

CBM's DSRKso2

IncompressibleNavierStokes.DSRK2 — Function

CBM's DSRK2

IncompressibleNavierStokes.DSRK3 — Function

Zennaro's DSRK3

"Non-SSP" Methods of Wong & Spiteri

Missing docstring.

Missing docstring for NSSP21. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NSSP32. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NSSP33. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NSSP53. Check Documenter's build log for details.