Pressure solvers

The discrete pressure Poisson equation

\[L p = W M F(u)\]

enforces divergence freeness. There are multiple options for solving this system.

IncompressibleNavierStokes.AbstractPressureSolver — Type

AbstractPressureSolver

Pressure solver for the Poisson equation.

IncompressibleNavierStokes.DirectPressureSolver — Type

DirectPressureSolver()

Direct pressure solver using a LU decomposition.

IncompressibleNavierStokes.CUDSSPressureSolver — Type

CUDSSPressureSolver()

Direct pressure solver using a CUDSS LDLt decomposition.

IncompressibleNavierStokes.CGPressureSolver — Type

CGPressureSolver(setup; [abstol], [reltol], [maxiter])

Conjugate gradients iterative pressure solver.

IncompressibleNavierStokes.CGMatrixPressureSolver — Type

CGMatrixPressureSolver(setup; [abstol], [reltol], [maxiter])

Conjugate gradients iterative pressure solver.

IncompressibleNavierStokes.SpectralPressureSolver — Type

SpectralPressureSolver(setup)

Build spectral pressure solver from setup.

IncompressibleNavierStokes.LowMemorySpectralPressureSolver — Type

LowMemorySpectralPressureSolver(setup)

Build spectral pressure solver from setup. This one is slower than SpectralPressureSolver but occupies less memory.

IncompressibleNavierStokes.pressure — Function

pressure(u, t, setup; psolver)

Do additional pressure solve. This makes the pressure compatible with the velocity field, resulting in same order pressure as velocity.

IncompressibleNavierStokes.pressure! — Function

pressure!(p, u, t, setup; psolver, F, div)

Compute pressure from velocity field. This makes the pressure compatible with the velocity field, resulting in same order pressure as velocity.

IncompressibleNavierStokes.poisson — Function

poisson(solver, f)

Solve the Poisson equation for the pressure with right hand side f at time t. For periodic and no-slip BC, the sum of f should be zero.

Non-mutating/allocating/out-of-place version.

See also poisson!.

IncompressibleNavierStokes.poisson! — Function

poisson!(solver, p, f)

Solve the Poisson equation for the pressure with right hand side f at time t. For periodic and no-slip BC, the sum of f should be zero.

Mutating/non-allocating/in-place version.

See also poisson.

IncompressibleNavierStokes.applypressure! — Function

applypressure!(u, p, setup)

Subtract pressure gradient (in-place).

IncompressibleNavierStokes.project — Function

project(u, setup; psolver)

Project velocity field onto divergence-free space.

IncompressibleNavierStokes.project! — Function

project!(u, setup; psolver, div, p)

Project velocity field onto divergence-free space.