IncompressibleNavierStokes

Documentation for IncompressibleNavierStokes.

IncompressibleNavierStokes.IncompressibleNavierStokes

IncompressibleNavierStokes.RungeKuttaMethod

IncompressibleNavierStokes.bodyforce

IncompressibleNavierStokes.bodyforce!

IncompressibleNavierStokes.build_operators!

IncompressibleNavierStokes.check_conservation

IncompressibleNavierStokes.check_input!

IncompressibleNavierStokes.check_symmetry

IncompressibleNavierStokes.convection!

IncompressibleNavierStokes.create_boundary_conditions!

IncompressibleNavierStokes.create_initial_conditions

IncompressibleNavierStokes.diffusion!

IncompressibleNavierStokes.filter_convection

IncompressibleNavierStokes.get_timestep

IncompressibleNavierStokes.interpolate_ν

IncompressibleNavierStokes.main

IncompressibleNavierStokes.momentum!

IncompressibleNavierStokes.momentum_rom

IncompressibleNavierStokes.nonuniform_grid

IncompressibleNavierStokes.operator_averaging!

IncompressibleNavierStokes.operator_divergence!

IncompressibleNavierStokes.operator_interpolation!

IncompressibleNavierStokes.plot_pressure

IncompressibleNavierStokes.plot_streamfunction

IncompressibleNavierStokes.plot_vorticity

IncompressibleNavierStokes.postprocess

IncompressibleNavierStokes.pressure_additional_solve!

IncompressibleNavierStokes.pressure_poisson

IncompressibleNavierStokes.set_bc_vectors!

IncompressibleNavierStokes.solve_steady!

IncompressibleNavierStokes.solve_steady_ibm!

IncompressibleNavierStokes.solve_steady_ke!

IncompressibleNavierStokes.solve_ubsteady_ke!

IncompressibleNavierStokes.solve_unsteady!

IncompressibleNavierStokes.solve_unsteady_rom!

IncompressibleNavierStokes.step_AB_CN!

IncompressibleNavierStokes.step_ERK!

IncompressibleNavierStokes.step_ERK_ROM

IncompressibleNavierStokes.strain_tensor

IncompressibleNavierStokes.tableau

IncompressibleNavierStokes.turbulent_viscosity

IncompressibleNavierStokes.IncompressibleNavierStokes — Module

Incompressible Navier-Stokes solvers

IncompressibleNavierStokes.RungeKuttaMethod — Type

RungeKuttaMethod

Abstract Runge Kutta method.

Original by David Ketcheson Extended by Benjamin Sanderse

IncompressibleNavierStokes.bodyforce — Function

Fx, Fy, dFx, dFy = force(V, t, setup, getJacobian) Body force in momentum equations in Finite Volume setting, so integrated dFx, dFy are the Jacobians dFx/dV and dFy/dV

IncompressibleNavierStokes.bodyforce! — Function

bodyforce!(F, ∇F V, t, setup, getJacobian)

Compute body force F in momentum equations at velocity points. If getJacobian, also compute ∇F = dF/dV.

If

IncompressibleNavierStokes.build_operators! — Method

build_operators!(setup)

Build discrete operators.

IncompressibleNavierStokes.check_conservation — Method

check_conservation(V, t, setup)

Check mass, momentum and energy conservation properties of velocity field

IncompressibleNavierStokes.check_input! — Method

Check input.

IncompressibleNavierStokes.check_symmetry — Function

Check symmetry of convection operator flag = 0: no symmetry error flag = 1: symmetry error

IncompressibleNavierStokes.convection! — Method

convection!(c, ∇c, V, ϕ, t, cache, setup, getJacobian) -> c, ∇c

evaluate convective terms and, optionally, Jacobians V: velocity field ϕ: "convection" field: e.g. d(ϕx u)/dx + d(ϕy u)/dy; usually ϕx = u, ϕy = v

IncompressibleNavierStokes.create_boundary_conditions! — Method

reate_boundary_conditions!(setup)

Create discrete boundary condtions.

IncompressibleNavierStokes.create_initial_conditions — Method

V, p = create_initial_conditions(setup)

Create initial vectors.

IncompressibleNavierStokes.diffusion! — Method

diffusion!(d, Jac, V, t, setup, getJacobian)

Evaluate diffusive terms and optionally Jacobian. Fill result in d. Return Jac.

IncompressibleNavierStokes.filter_convection — Method

filter_convection(u, diff_matrix, bc, α)

Construct filter for convective terms

IncompressibleNavierStokes.get_timestep — Method

estimate time step based on eigenvalues of operators, using Gershgorin

IncompressibleNavierStokes.interpolate_ν — Method

ν_t_ux, ν_t_uy, ν_t_vx, ν_t_vy = interpolate_ν(ν_t, setup)

interpolate the scalar field νt at pressure locations (xp, yp) to locations needed in computing the diffusive terms, i.e. the ux, uy, vx and v_y locations

IncompressibleNavierStokes.main — Method

This m-file contains the code for the 2D incompressible Navier-Stokes equations using a Finite Volume Method and a pressure correction method.

horizontal numbering of volumes

2nd and 4th order spatial (central) discretization convection and diffusion

general boundary conditions different fourth order BC ("verstappen") can be used by changing

addpath("spatial/boundaryconditions/proposed") to addpath("spatial/boundaryconditions/verstappen") check operatorconvectiondiffusion construction of Duy and Dvx see readme.txt Benjamin Sanderse, September 2018 - April 2019

IncompressibleNavierStokes.momentum! — Function

momentum!(F, ∇F, V, ϕ, p, t, setup, cache = MomentumCache(setup), getJacobian = false, nopressure = false)

Calculate rhs of momentum equations and, optionally, Jacobian with respect to velocity field V: velocity field ϕ: "convection" field: e.g. d(cx u)/dx + d(cy u)/dy; usually cx = u, cy = v, so ϕ = V p: pressure getJacobian = true: return ∇FdV nopressure = true: exclude pressure gradient; in this case input argument p is not used

IncompressibleNavierStokes.momentum_rom — Method

momentum_rom()

Rhs momentum equation for ROM.

IncompressibleNavierStokes.nonuniform_grid — Function

Cenerate a non-uniform grid, from zlow to zup, starting deltaz and having stretch factor close to sz v2.0: includes check for uniform grid and, in that case, adapts deltaz if necessary

IncompressibleNavierStokes.operator_averaging! — Method

Construct averaging operators.

IncompressibleNavierStokes.operator_divergence! — Method

Construct divergence and gradient operator

IncompressibleNavierStokes.operator_interpolation! — Method

Construct averaging operators.

IncompressibleNavierStokes.plot_pressure — Method

Plot pressure.

IncompressibleNavierStokes.plot_streamfunction — Method

Plot streamfunction.

IncompressibleNavierStokes.plot_vorticity — Method

plot_vorticity.

IncompressibleNavierStokes.postprocess — Method

Postprocess.

IncompressibleNavierStokes.pressure_additional_solve! — Method

Additional pressure solve. make the pressure compatible with the velocity field. this should also result in same order pressure as velocity

IncompressibleNavierStokes.pressure_poisson — Function

Solve the Poisson equation for the pressure. compute pressure from pressure poisson_solver problem with right-hand side f assume the Laplace matrix is known (A) and is possibly factorized (LU); right hand side is given by f we should have sum(f) = 0 for periodic and no-slip BC

IncompressibleNavierStokes.set_bc_vectors! — Method

set_bc_vectors!(setup, t)

Construct boundary conditions.

IncompressibleNavierStokes.solve_steady! — Method

solve_steady!(solution, setup)

Solve the entire saddlepoint system arising from the steady Navier-Stokes equations with linearization of the convective terms

IncompressibleNavierStokes.solve_steady_ibm! — Method

solve_steady_ibm!(solution, setup)

Solve ibm.

IncompressibleNavierStokes.solve_steady_ke! — Method

solve_steady_ke!(solution, setup)

Solve ke.

IncompressibleNavierStokes.solve_ubsteady_ke! — Method

solve_unsteady_ke!(solution, setup)

Solve ke.

IncompressibleNavierStokes.solve_unsteady! — Method

solve_unsteady!(solution, setup)

Main solver file for unsteady calculations

IncompressibleNavierStokes.solve_unsteady_rom! — Method

solve_unsteady_rom!(solution, setup)

Solve ke.

IncompressibleNavierStokes.step_AB_CN! — Method

convₙ₋₁ are the convection terms of t^(n-1) output includes convection terms at t^(n), which will be used in next time step in the Adams-Bashforth part of the method

Adams-Bashforth for convection and Crank-Nicolson for diffusion formulation: (u^{n+1} - u^{n})/Δt = -(α₁(conv^n) + α₂(conv^{n-1})) + θdiff^{n+1} + (1-θ)diff^{n} + θF^{n+1} + (1-θ)F^{n} θBC^{n+1} + (1-θ)BC^{n} - G*p + y_p where BC are boundary conditions of diffusion

rewrite as: (I/Δt - θD)u^{n+1} = (I/Δt - (1-θ)D)u^{n} + -(α₁(conv^n) + α₂(conv^{n-1})) + θF^{n+1} + (1-θ)F^{n} θBC^{n+1} + (1-θ)BC^{n} - G*p + y_p

the LU decomposition of the first matrix is precomputed in operatorconvectiondiffusion

note that, in constrast to explicit methods, the pressure from previous time steps has an influence on the accuracy of the velocity

IncompressibleNavierStokes.step_ERK! — Method

step_ERK()

Perform one time step for the general explicit Runge-Kutta method (ERK).

Dirichlet boundary points are not part of solution vector but are prescribed in a strong manner via the ubc and vbc functions.

IncompressibleNavierStokes.step_ERK_ROM — Method

step_ERK_ROM()

General explicit Runge-Kutta method for ROM Perform one tᵢme step for the general explicit Runge-Kutta method (ERK) with Reduced Order Model (ROM).

IncompressibleNavierStokes.strain_tensor — Method

S11, S12, S21, S22, S_abs, Jacu, Jacv = strain_tensor(V, t, setup, getJacobian)

Evaluate rate of strain tensor S(u) and its magnitude

IncompressibleNavierStokes.tableau — Function

A, b, c, r = tableau(rk_method, s = 1)

Set up Butcher arrays A, b, and c for the given rk_method. Also return SSP coefficient r. For families of methods, optional input s is the number of stages.

IncompressibleNavierStokes.turbulent_viscosity — Method

nu_t = turbulent_viscosity(S_abs, setup)

Compute turbulent viscosity based on S_abs