IncompressibleNavierStokes

Documentation for IncompressibleNavierStokes.

IncompressibleNavierStokes.IncompressibleNavierStokes

IncompressibleNavierStokes.AdamsBashforthCrankNicolsonStepper

IncompressibleNavierStokes.MomentumCache

IncompressibleNavierStokes.OneLegStepper

IncompressibleNavierStokes.PressureSolver

IncompressibleNavierStokes.RungeKuttaStepper

IncompressibleNavierStokes.TimeStepper

IncompressibleNavierStokes.TimeStepperCache

IncompressibleNavierStokes.bodyforce

IncompressibleNavierStokes.bodyforce!

IncompressibleNavierStokes.build_operators!

IncompressibleNavierStokes.check_input!

IncompressibleNavierStokes.check_symmetry

IncompressibleNavierStokes.compute_conservation

IncompressibleNavierStokes.convection

IncompressibleNavierStokes.convection!

IncompressibleNavierStokes.convection_components!

IncompressibleNavierStokes.create_boundary_conditions!

IncompressibleNavierStokes.create_initial_conditions

IncompressibleNavierStokes.diffusion!

IncompressibleNavierStokes.filter_convection

IncompressibleNavierStokes.filter_convection!

IncompressibleNavierStokes.get_timestep

IncompressibleNavierStokes.get_velocity

IncompressibleNavierStokes.get_vorticity

IncompressibleNavierStokes.initialize!

IncompressibleNavierStokes.initialize_rtp

IncompressibleNavierStokes.interpolate_ν

IncompressibleNavierStokes.is_steady

IncompressibleNavierStokes.main

IncompressibleNavierStokes.momentum

IncompressibleNavierStokes.momentum!

IncompressibleNavierStokes.momentum_allstage!

IncompressibleNavierStokes.momentum_rom

IncompressibleNavierStokes.nonuniform_grid

IncompressibleNavierStokes.nstage

IncompressibleNavierStokes.operator_averaging!

IncompressibleNavierStokes.operator_convection_diffusion!

IncompressibleNavierStokes.operator_divergence!

IncompressibleNavierStokes.operator_interpolation!

IncompressibleNavierStokes.operator_postprocessing!

IncompressibleNavierStokes.operator_turbulent_diffusion!

IncompressibleNavierStokes.operator_viscosity!

IncompressibleNavierStokes.plot_pressure

IncompressibleNavierStokes.plot_streamfunction

IncompressibleNavierStokes.plot_vorticity

IncompressibleNavierStokes.postprocess

IncompressibleNavierStokes.pressure_additional_solve!

IncompressibleNavierStokes.pressure_poisson!

IncompressibleNavierStokes.set_bc_vectors!

IncompressibleNavierStokes.solve

IncompressibleNavierStokes.solve

IncompressibleNavierStokes.solve

IncompressibleNavierStokes.solve_steady_ibm!

IncompressibleNavierStokes.solve_steady_ke!

IncompressibleNavierStokes.solve_ubsteady_ke!

IncompressibleNavierStokes.solve_unsteady_rom!

IncompressibleNavierStokes.step!

IncompressibleNavierStokes.step!

IncompressibleNavierStokes.step!

IncompressibleNavierStokes.step!

IncompressibleNavierStokes.step!

IncompressibleNavierStokes.step_erk_rom

IncompressibleNavierStokes.strain_tensor

IncompressibleNavierStokes.tableau

IncompressibleNavierStokes.time_stepper_cache

IncompressibleNavierStokes.turbulent_viscosity

IncompressibleNavierStokes.update_rtp!

IncompressibleNavierStokes.vorticity!

IncompressibleNavierStokes.IncompressibleNavierStokes — Module

Incompressible Navier-Stokes solvers

IncompressibleNavierStokes.AdamsBashforthCrankNicolsonStepper — Type

AdamsBashforthCrankNicolsonStepper(; α₁ = 3 // 2, α₂ = -1 // 2, θ = 1 // 2)

Adams-Bashforth for velocity (parameters α₁ and α₂) and Crank-Nicolson for pressure (with implicitness θ).

IncompressibleNavierStokes.MomentumCache — Type

MomentumCache

Preallocation structure for terms in the momentum equations.

IncompressibleNavierStokes.OneLegStepper — Type

OneLegStepper(β = 1 // 2)

One-leg β-method.

IncompressibleNavierStokes.PressureSolver — Type

PressureSolver

Lineare pressure solver for the Poisson equation.

IncompressibleNavierStokes.RungeKuttaStepper — Type

RungeKuttaStepper

Abstract Runge Kutta time stepper.

IncompressibleNavierStokes.TimeStepper — Type

TimeStepper

Abstract time stepper.

IncompressibleNavierStokes.TimeStepperCache — Type

TimeStepperCache

Time stepper cache.

IncompressibleNavierStokes.bodyforce! — Method

bodyforce!(F, ∇F V, t, setup; getJacobian = false)

Compute body force F in momentum equations at velocity points. If getJacobian, also compute ∇F = ∂F/∂V.

IncompressibleNavierStokes.bodyforce — Method

bodyforce(V, t, setup; getJacobian = false)

Body force in momentum equations in Finite Volume setting, so integrated dFx, dFy are the Jacobians ∂Fx/∂V and ∂Fy/∂V.

IncompressibleNavierStokes.build_operators! — Method

build_operators!(setup)

Build discrete operators.

IncompressibleNavierStokes.check_input! — Method

Check input.

IncompressibleNavierStokes.check_symmetry — Function

check_symmetry(V, t, setup, ϵ = 1e-14)

Check symmetry of convection operator.

flag = 0: no symmetry error flag = 1: symmetry error

IncompressibleNavierStokes.compute_conservation — Method

compute_conservation(V, t, setup)

Compute mass, momentum and energy conservation properties of velocity field.

IncompressibleNavierStokes.convection! — Method

convection!(c, ∇c, V, ϕ, t, cache, setup, getJacobian = false) -> c, ∇c

evaluate convective terms c and, optionally, Jacobian ∇c = ∂c/∂V.

IncompressibleNavierStokes.convection — Method

convection(V, ϕ, t, setup, getJacobian = false)

Convenience function for initializing arrays c and ∇c before filling in convection terms.

IncompressibleNavierStokes.convection_components! — Method

convection_components!(c, ∇c, V, ϕ, setup, cache; getJacobian = false, order4 = false)

Compute convection components.

IncompressibleNavierStokes.create_boundary_conditions! — Method

reate_boundary_conditions!(setup)

Create discrete boundary condtions.

Values should either be scalars or vectors. All values (u, v, p, k, e) are defined at x, y locations, i.e. the corners of pressure volumes, so they cover the entire domain, including corners.

IncompressibleNavierStokes.create_initial_conditions — Method

V, p = create_initial_conditions(setup)

Create initial vectors.

IncompressibleNavierStokes.diffusion! — Method

diffusion!(d, ∇d, V, t, setup; getJacobian = false)

Evaluate diffusive terms d and optionally Jacobian ∇d = ∂d/∂V.

IncompressibleNavierStokes.filter_convection! — Method

filter_convection!(ū, u, diff_matrix, bc, α)

Construct filter for convective terms.

IncompressibleNavierStokes.filter_convection — Method

filter_convection(u, diff_matrix, bc, α)

Construct filter for convective terms.

IncompressibleNavierStokes.get_timestep — Method

get_timestep(setup)

Estimate time step based on eigenvalues of operators, using Gershgorin.

IncompressibleNavierStokes.get_velocity — Method

get_velocity(V, t, setup)

Get velocity values at pressure points. Interpolate velocities to pressure positions using BMx and BMy (constructed in operator_divergence.jl).

IncompressibleNavierStokes.get_vorticity — Method

get_vorticity(V, t, setup)

Get vorticity from velocity field.

IncompressibleNavierStokes.initialize! — Function

initialize(pressure_solver)

Initialize pressure solver.

IncompressibleNavierStokes.initialize_rtp — Method

initialize_rtp(setup, V, p, t)

Initialize real time plot (RTP).

IncompressibleNavierStokes.interpolate_ν — Method

ν_t_ux, ν_t_uy, ν_t_vx, ν_t_vy = interpolate_ν(ν_t, setup)

interpolate the scalar field νt at pressure locations (xp, yp) to locations needed in computing the diffusive terms, i.e. the ux, uy, vx and v_y locations

IncompressibleNavierStokes.is_steady — Function

is_steady(problem)

Returns true if problem is unsteady.

IncompressibleNavierStokes.main — Method

main(setup)

Build mesh and operators from setup, solve problem, and postprocess. Return solution and timing.

IncompressibleNavierStokes.momentum! — Method

momentum!(F, ∇F, V, ϕ, p, t, setup, cache; getJacobian = false, nopressure = false)

Calculate rhs of momentum equations and, optionally, Jacobian with respect to velocity field.

V: velocity field ϕ: "convection" field: e.g. ∂(ϕx V)/∂x + ∂(ϕy V)/∂y; usually ϕ = V (so ϕx = u, ϕy = v) p: pressure getJacobian: return ∇F = ∂F/∂V nopressure: exclude pressure gradient; in this case input argument p is not used

IncompressibleNavierStokes.momentum — Method

momentum(V, ϕ, p, t, setup; getJacobian = false, nopressure = false)

Convenience function for initializing arrays F and ∇F before filling in momentum terms.

IncompressibleNavierStokes.momentum_allstage! — Method

momentum_allstage!(F, ∇F, V, C, p, t, setup, stepper_cache, momentum_cache; getJacobian = false)

Call momentum for multiple (V, p) pairs, as required in implicit RK methods.

IncompressibleNavierStokes.momentum_rom — Method

momentum_rom()

Rhs momentum equation for ROM.

IncompressibleNavierStokes.nonuniform_grid — Function

Cenerate a non-uniform grid, from z_low to z_up, starting deltaz and having stretch factor close to sz Includes check for uniform grid and, in that case, adapts deltaz if necessary.

IncompressibleNavierStokes.nstage — Method

nstage(rk_stepper)

Get number of stages of the Runge-Kutta time stepper rk_stepper.

IncompressibleNavierStokes.operator_averaging! — Method

Construct averaging operators.

IncompressibleNavierStokes.operator_convection_diffusion! — Method

Construct convection and diffusion operators.

IncompressibleNavierStokes.operator_divergence! — Method

Construct divergence and gradient operator

IncompressibleNavierStokes.operator_interpolation! — Method

Construct averaging operators.

IncompressibleNavierStokes.operator_postprocessing! — Method

operator_postprocessing!(setup)

Construct postprocessing operators such as vorticity.

IncompressibleNavierStokes.operator_turbulent_diffusion! — Method

operator_turbulent_diffusion!(setup)

Average (turbulent) viscosity to cell faces: from ν at xp, yp to ν at ux, uy, vx, vy locations

See also ke_viscosity.jl.

IncompressibleNavierStokes.operator_viscosity! — Function

operator_viscosity!(model, setup)

Classical turbulence modelling via the diffusive term

IncompressibleNavierStokes.plot_pressure — Method

Plot pressure.

IncompressibleNavierStokes.plot_streamfunction — Method

Plot streamfunction.

IncompressibleNavierStokes.plot_vorticity — Method

plot_vorticity.

IncompressibleNavierStokes.postprocess — Method

Postprocess.

IncompressibleNavierStokes.pressure_additional_solve! — Method

Additional pressure solve. make the pressure compatible with the velocity field. this should also result in same order pressure as velocity

IncompressibleNavierStokes.pressure_poisson! — Function

pressure_poisson!(solver, p, f, t, setup, tol = 1e-14)

Solve the Poisson equation for the pressure with right hand side f at time t.

We should have sum(f) = 0 for periodic and no-slip BC.

IncompressibleNavierStokes.set_bc_vectors! — Method

set_bc_vectors!(setup, t)

Construct boundary conditions.

IncompressibleNavierStokes.solve — Function

solve(problem, setup, V₀, p₀)

Solve problem with initial state (V₀, p₀).

IncompressibleNavierStokes.solve — Method

solve(steady_state_problem, setup, V₀, p₀)

Solve steady state problem of the Navier-Stokes equations. This saddlepoint system arises from linearization of the convective terms.

IncompressibleNavierStokes.solve — Method

solve(unsteady_problem, setup, V₀, p₀)

Solve unsteady_problem.

IncompressibleNavierStokes.solve_steady_ibm! — Method

solve_steady_ibm!(solution, setup)

Solve ibm.

IncompressibleNavierStokes.solve_steady_ke! — Method

solve_steady_ke!(solution, setup)

Solve ke.

IncompressibleNavierStokes.solve_ubsteady_ke! — Method

solve_unsteady_ke!(solution, setup)

Solve ke.

IncompressibleNavierStokes.solve_unsteady_rom! — Method

solve_unsteady_rom!(solution, setup)

Solve ke.

IncompressibleNavierStokes.step! — Function

step!(time_stepper, args...; kwargs...)

Do one time step with a timestepper.

IncompressibleNavierStokes.step! — Method

step!(
    ts::AdamsBashforthCrankNicolsonStepper,
    V,
    p,
    Vₙ,
    pₙ,
    Vₙ₋₁,
    pₙ₋₁,
    cₙ₋₁,
    tₙ,
    Δt,
    setup,
    stepper_cache,
    momentum_cache,
)

Perform one time step with Adams-Bashforth for convection and Crank-Nicolson for diffusion.

cₙ₋₁ are the convection terms of tₙ₋₁. Output includes convection terms at tₙ, which will be used in next time step in the Adams-Bashforth part of the method.

Adams-Bashforth for convection and Crank-Nicolson for diffusion formulation:

(u^{n+1} - u^{n})/Δt = -(α₁ c^n + α₂ c^{n-1}) + θ diff^{n+1} + (1-θ) diff^{n} + θ F^{n+1} + (1-θ) F^{n} + θ BC^{n+1} + (1-θ) BC^{n} - G*p + y_p

where BC are boundary conditions of diffusion. This is rewritten as:

(I/Δt - θ D) u^{n+1} = (I/Δt - (1-θ) D) u^{n} - (α₁ c^n + α₂ c^{n-1}) + θ F^{n+1} + (1-θ) F^{n} + θ BC^{n+1} + (1-θ) BC^{n} - G*p + y_p

The LU decomposition of the LHS matrix is precomputed in operator_convection_diffusion.jl.

Note that, in constrast to explicit methods, the pressure from previous time steps has an influence on the accuracy of the velocity.

IncompressibleNavierStokes.step! — Method

step!(ts::ExplicitRungeKuttaStepper, V, p, Vₙ, pₙ, Vₙ₋₁, pₙ₋₁, tₙ, Δtₙ, setup, stepper_cache, momentum_cache)

Perform one time step for the general explicit Runge-Kutta method (ERK).

Dirichlet boundary points are not part of solution vector but are prescribed in a strong manner via the u_bc and v_bc functions.

IncompressibleNavierStokes.step! — Method

step!(ts::ImplicitRungeKuttaStepper, V, p, Vₙ, pₙ, Vₙ₋₁, pₙ₋₁, tₙ, Δtₙ, setup, momentum_cache)

Do one time step for implicit Runge-Kutta method.

Unsteady Dirichlet boundary points are not part of solution vector but are prescribed in a "strong" manner via the u_bc and v_bc functions.

IncompressibleNavierStokes.step! — Method

step!(ts::OneLegStepper, V, p, Vₙ, pₙ, Vₙ₋₁, pₙ₋₁, tₙ, Δtₙ, setup, momentum_cache)

Do one time step using one-leg-β-method following symmetry-preserving discretization of turbulent flow. See [Verstappen and Veldman (JCP 2003)] for details, or [Direct numerical simulation of turbulence at lower costs (Journal of Engineering Mathematics 1997)].

Formulation: $\frac{(\beta + 1/2) u^{n+1} - 2 \beta u^{n} + (\beta - 1/2) u^{n-1}}{\Delta t} = F((1 + \beta) u^n - \beta u^{n-1})$

IncompressibleNavierStokes.step_erk_rom — Method

step_ERK_ROM()

Perform one time step for the general explicit Runge-Kutta method (ERK) with Reduced Order Model (ROM).

IncompressibleNavierStokes.strain_tensor — Method

strain_tensor(V, t, setup; getJacobian = false, get_S_abs = false)

Evaluate rate of strain tensor S(V) and its magnitude.

IncompressibleNavierStokes.tableau — Function

A, b, c, r = tableau(rk_stepper, s = 1)

Set up Butcher arrays A, b, and c for the given rk_stepper. Also return SSP coefficient r. For families of methods, optional input s is the number of stages.

Original (MATLAB) by David Ketcheson, extended by Benjamin Sanderse.

IncompressibleNavierStokes.time_stepper_cache — Function

time_stepper_cache(stepper, args...; kwargs...)

Get time stepper cache for the given time stepper.

IncompressibleNavierStokes.turbulent_viscosity — Function

turbulent_viscosity(model, S_abs)

Compute turbulent viscosity based on S_abs.

IncompressibleNavierStokes.update_rtp! — Method

update_rtp!(rtp, setup, V, p, t)

IncompressibleNavierStokes.vorticity! — Method

vorticity!(ω, V, t, setup)

Compute vorticity values at pressure midpoints. This should be consistent with operator_postprocessing.jl.