API Reference

IncompressibleNavierStokes.IncompressibleNavierStokes — Module

IncompressibleNavierStokes

Energy-conserving solvers for the incompressible Navier-Stokes equations.

IncompressibleNavierStokes.Setup — Function

Setup(
    x...;
    boundary_conditions = ntuple(d -> (PeriodicBC(), PeriodicBC()), length(x)),
    Re = convert(eltype(x[1]), 1_000),
    viscosity_model = LaminarModel(),
    convection_model = NoRegConvectionModel(),
    bodyforce = nothing,
    closure_model = nothing,
    ArrayType = Array,
)

Create setup.

Grid

IncompressibleNavierStokes.Dimension — Type

Dimension(N)

Represent an N-dimensional space. Returns N when called.

julia> d = Dimension(3)
Dimension{3}()

julia> d()
3

IncompressibleNavierStokes.Grid — Function

Grid(x, boundary_conditions)

Create nonuniform Cartesian box mesh x[1] × ... × x[d] with boundary conditions boundary_conditions.

IncompressibleNavierStokes.cosine_grid — Function

cosine_grid(a, b, N)

Create a nonuniform grid of N + 1 points from a to b using a cosine profile, i.e.

\[x_i = a + \frac{1}{2} \left( 1 - \cos \left( \pi \frac{i}{n} \right) \right) (b - a), \quad i = 0, \dots, N\]

See also stretched_grid.

IncompressibleNavierStokes.max_size — Function

max_size(grid)

Get size of the largest grid element.

IncompressibleNavierStokes.stretched_grid — Function

stretched_grid(a, b, N, s = 1)

Create a nonuniform grid of N + 1 points from a to b with a stretch factor of s. If s = 1, return a uniform spacing from a to b. Otherwise, return a vector $x \in \mathbb{R}^{N + 1}$ such that $x_n = a + \sum_{i = 1}^n s^{i - 1} h$ for $n = 0, \dots , N$. Setting $x_N = b$ then gives $h = (b - a) \frac{1 - s}{1 - s^N}$, resulting in

\[x_n = a + (b - a) \frac{1 - s^n}{1 - s^N}, \quad n = 0, \dots, N.\]

Note that stretched_grid(a, b, N, s)[n] corresponds to $x_{n - 1}$.

See also cosine_grid.

Visocosity Models

IncompressibleNavierStokes.AbstractViscosityModel — Type

AbstractViscosityModel

Abstract viscosity model.

IncompressibleNavierStokes.LaminarModel — Type

LaminarModel()

Laminar model.

IncompressibleNavierStokes.MixingLengthModel — Type

MixingLengthModel()

Mixing-length model with mixing length lm.

IncompressibleNavierStokes.SmagorinskyModel — Type

SmagorinskyModel(C_s = 0.17)

Smagorinsky-Lilly model with constant C_s.

IncompressibleNavierStokes.QRModel — Type

QR(Re)

QR-model.

Convection Models

IncompressibleNavierStokes.AbstractConvectionModel — Type

AbstractConvectionModel

Abstract convection model.

IncompressibleNavierStokes.NoRegConvectionModel — Type

NoRegConvection()

Unregularized convection model.

IncompressibleNavierStokes.C2ConvectionModel — Type

C2ConvectionModel()

C2 regularization convection model.

IncompressibleNavierStokes.C4ConvectionModel — Type

C4ConvectionModel()

C4 regularization convection model.

IncompressibleNavierStokes.LerayConvectionModel — Type

LerayConvectionModel()

Leray regularization convection model.

Momentum

IncompressibleNavierStokes.divergence — Function

divergence(u, setup)

Compute divergence of velocity field.

IncompressibleNavierStokes.divergence! — Function

divergence!(M, u, setup)

Compute divergence of velocity field (in-place version).

IncompressibleNavierStokes.vorticity — Function

vorticity(u, setup)

Compute vorticity field.

IncompressibleNavierStokes.vorticity! — Function

vorticity!(ω, u, setup)

Compute vorticity field.

IncompressibleNavierStokes.convection! — Function

convection!(F, u, setup)

Compute convective term.

IncompressibleNavierStokes.diffusion! — Function

diffusion!(F, u, setup)

Compute diffusive term.

IncompressibleNavierStokes.bodyforce! — Function

bodyforce!(F, u, setup)

Compute body force.

IncompressibleNavierStokes.momentum! — Function

momentum!(F, u, t, setup)

Right hand side of momentum equations, excluding pressure gradient. Put the result in $F$.

Postprocess

IncompressibleNavierStokes.plot_force — Function

plot_force(setup, t; kwargs...)

Plot body force.

IncompressibleNavierStokes.plot_grid — Function

plot_grid(x, y)
plot_grid(x, y, z)
plot_grid(grid)

Plot nonuniform Cartesian grid.

IncompressibleNavierStokes.plot_pressure — Function

plot_pressure(setup, p; kwargs...)

Plot pressure.

IncompressibleNavierStokes.plot_streamfunction — Function

plot_streamfunction(setup, V, t; kwargs...)

Plot streamfunction.

IncompressibleNavierStokes.plot_velocity — Function

plot_velocity(setup, V, t; kwargs...)

Plot velocity.

IncompressibleNavierStokes.plot_vorticity — Function

plot_vorticity(setup, V, t; kwargs...)

Plot vorticity field.

IncompressibleNavierStokes.save_vtk — Function

save_vtk(setup, u, p, filename = "output/solution")

Save velocity and pressure field to a VTK file.

In the case of a 2D setup, the velocity field is saved as a 3D vector with a z-component of zero, as this seems to be preferred by ParaView.

Preprocess

IncompressibleNavierStokes.create_initial_conditions — Function

create_initial_conditions(
    setup,
    initial_velocity,
    t = 0;
    pressure_solver = DirectPressureSolver(setup),
)

Create initial vectors (u, p) at starting time t.

IncompressibleNavierStokes.random_field — Function

random_field(
    setup, t;
    A = setup.grid.N[1] * 10_000,
    σ = 30,
    s = 5,
    pressure_solver = DirectPressureSolver(setup),
)

Create random field.

K: Maximum wavenumber

A: Eddy amplitude

σ: Variance

s Wavenumber offset before energy starts decaying

Processors

IncompressibleNavierStokes.step_logger — Function

step_logger(; nupdate = 1)

Create processor that logs time step information.

IncompressibleNavierStokes.vtk_writer — Function

vtk_writer(
    setup;
    nupdate = 1,
    dir = "output",
    filename = "solution",
    fields = (:velocity, :pressure, :vorticity),
)

Create processor that writes the solution every nupdate time steps to a VTK file. The resulting Paraview data collection file is stored in "$dir/$filename.pvd".

IncompressibleNavierStokes.field_saver — Function

field_saver(setup; nupdate = 1)

Create processor that stores the solution every nupdate time step to the vector of vectors V and p. The solution times are stored in the vector t.

IncompressibleNavierStokes.field_plotter — Function

field_plotter(
    setup;
    fieldname = :vorticity,
    type = nothing,
    sleeptime = 0.001,
    alpha = 0.05,
)

Plot the solution every time the state o is updated.

The sleeptime is slept at every update, to give Makie time to update the plot. Set this to nothing to skip sleeping.

Available fieldnames are:

:velocity,

:vorticity,

:streamfunction,

:pressure.

Available plot types for 2D are:

heatmap (default),

image,

contour,

contourf.

Available plot types for 3D are:

contour (default).

The alpha value gets passed to contour in 3D.

IncompressibleNavierStokes.energy_history_plotter — Function

energy_history_plotter(setup)

Create energy history plot, with a history point added every time step_observer is updated.

IncompressibleNavierStokes.energy_spectrum_plotter — Function

energy_spectrum_plotter(setup; nupdate = 1)

Create energy spectrum plot, redrawn every time step_observer is updated.

IncompressibleNavierStokes.animator — Function

animator(setup, path; nupdate = 1, plotter = field_plotter(setup); kwargs...)

Animate a plot of the solution every update iteration. The animation is saved to path, which should have one of the following extensions:

".mkv"

".mp4"

".webm"

".gif"

The plot is determined by a plotter processsor. Addtional kwargs are passed to Makie's VideoStream.

Solvers

IncompressibleNavierStokes.get_timestep — Function

get_timestep(stepper, cfl; bc_vectors)

Estimate time step based on eigenvalues of operators, using Gershgorin.

IncompressibleNavierStokes.solve_unsteady — Function

function solve_unsteady(
    setup, u₀, p₀, tlims;
    method = RK44(; T = eltype(u₀)),
    pressure_solver = DirectPressureSolver(setup),
    Δt = zero(eltype(u₀)),
    cfl = 1,
    n_adapt_Δt = 1,
    inplace = true,
    processors = (),
)

Solve unsteady problem using method.

If Δt is a real number, it is rounded such that (t_end - t_start) / Δt is an integer. If Δt = nothing, the time step is chosen every n_adapt_Δt iteration with CFL-number cfl .

Each processor is called after every processor.nupdate time step.

All arrays and operators are passed through the device function. This allows for performing computations on a different device than the host (CPU). To compute on an Nvidia GPU using CUDA, change

solve_unsteady(setup, u₀, p₀, tlims; kwargs...)

to the following:

using CUDA
solve_unsteady(
    setup, u₀, p₀, tlims;
    device = cu,
    kwargs...
)

Note that the state observable passed to the processor.initialize function contains vector living on the device, and you may have to move them back to the host using Array(u) and Array(p) in the processor.

IncompressibleNavierStokes.solve_steady_state — Function

function solve_steady_state(
    setup, V₀, p₀;
    jacobian_type = :newton,
    npicard = 2,
    abstol = 1e-10,
    maxiter = 10,
)

Solve steady state problem of the Navier-Stokes equations. This saddlepoint system arises from linearization of the convective terms.

Each processor is called after every processor.nupdate iteration.

Pressure solvers

IncompressibleNavierStokes.AbstractPressureSolver — Type

AbstractPressureSolver

Pressure solver for the Poisson equation.

IncompressibleNavierStokes.DirectPressureSolver — Type

DirectPressureSolver()

Direct pressure solver using a LU decomposition.

IncompressibleNavierStokes.CGPressureSolver — Type

CGPressureSolver(setup; [abstol], [reltol], [maxiter])

Conjugate gradients iterative pressure solver.

IncompressibleNavierStokes.CGPressureSolverManual — Type

CGPressureSolverManual(setup; [abstol], [reltol], [maxiter])

Conjugate gradients iterative pressure solver.

IncompressibleNavierStokes.SpectralPressureSolver — Type

SpectralPressureSolver(setup)

Build spectral pressure solver from setup.

IncompressibleNavierStokes.pressure_additional_solve — Function

pressure_additional_solve(pressure_solver, u, t, setup)

Do additional pressure solve. This makes the pressure compatible with the velocity field, resulting in same order pressure as velocity.

IncompressibleNavierStokes.pressure_additional_solve! — Function

pressure_additional_solve!(pressure_solver, u, p, t, setup, F, f, M)

Do additional pressure solve. This makes the pressure compatible with the velocity field, resulting in same order pressure as velocity.

IncompressibleNavierStokes.pressure_poisson — Function

pressure_poisson(solver, f)

Solve the Poisson equation for the pressure with right hand side f at time t. For periodic and no-slip BC, the sum of f should be zero.

Non-mutating/allocating/out-of-place version.

See also pressure_poisson!.

IncompressibleNavierStokes.pressure_poisson! — Function

pressure_poisson!(solver, p, f)

Solve the Poisson equation for the pressure with right hand side f at time t. For periodic and no-slip BC, the sum of f should be zero.

Mutating/non-allocating/in-place version.

See also pressure_poisson.

Time steppers

IncompressibleNavierStokes.AbstractODEMethod — Type

AbstractODEMethod

Abstract ODE method.

IncompressibleNavierStokes.AbstractRungeKuttaMethod — Type

AbstractRungeKuttaMethod

Abstract Runge Kutta method.

IncompressibleNavierStokes.AdamsBashforthCrankNicolsonMethod — Type

AdamsBashforthCrankNicolsonMethod(
    T = Float64;
    α₁ = T(3 // 2),
    α₂ = T(-1 // 2),
    θ = T(1 // 2),
    p_add_solve = true,
    method_startup = RK44(; T),
)

IMEX AB-CN: Adams-Bashforth for explicit convection (parameters α₁ and α₂) and Crank-Nicolson for implicit diffusion (implicitness θ). The method is second order for θ = 1/2.

The LU decomposition of the LHS matrix is computed every time the time step changes.

Note that, in constrast to explicit methods, the pressure from previous time steps has an influence on the accuracy of the velocity.

IncompressibleNavierStokes.OneLegMethod — Type

OneLegMethod(
    T = Float64;
    β = T(1 // 2),
    p_add_solve = true,
    method_startup = RK44(; T),
)

Explicit one-leg β-method following symmetry-preserving discretization of turbulent flow. See Verstappen and Veldman [5] [6] for details.

IncompressibleNavierStokes.ExplicitRungeKuttaMethod — Type

ExplicitRungeKuttaMethod(; A, b, c, r, p_add_solve = true)

Explicit Runge Kutta method.

IncompressibleNavierStokes.ImplicitRungeKuttaMethod — Type

ImplicitRungeKuttaMethod(;
    A,
    b,
    c,
    r,
    newton_type = :full,
    maxiter = 10,
    abstol = 1e-14,
    reltol = 1e-14,
    p_add_solve = true,
)

Implicit Runge Kutta method.

The implicit linear system is solved at each time step using Newton's method. The newton_type may be one of the following:

:no: Replace iteration matrix with I/Δt (no Jacobian)

:approximate: Build Jacobian once before iterations only

:full: Build Jacobian at each iteration

IncompressibleNavierStokes.isexplicit — Function

isexplicit(method)

Return true if method is explicit, i.e. the value at a certain time step is given explicitly as a function of the previous time steps only.

IncompressibleNavierStokes.lambda_conv_max — Function

lambda_conv_max(method)

Get maximum value of stability region for the convection operator (not a very good indication for the methods that do not include the imaginary axis)

IncompressibleNavierStokes.lambda_diff_max — Function

lambda_diff_max(method)

Get maximum value of stability region for the diffusion operator.

IncompressibleNavierStokes.nstage — Function

nstage(rk_method)

Get number of stages of the Runge-Kutta method.

IncompressibleNavierStokes.ode_method_cache — Function

ode_method_cache(method, setup)

Get time stepper cache for the given ODE method.

IncompressibleNavierStokes.runge_kutta_method — Function

runge_kutta_method(A, b, c, r; [p_add_solve], [newton_type], [maxiter], [abstol], [reltol])

Get Runge Kutta method. The function checks whether the method is explicit.

p_add_solve: whether to add a pressure solve step to the method.

For implicit RK methods: newton_type, maxiter, abstol, reltol.

IncompressibleNavierStokes.timestep — Function

step(stepper, Δt; bc_vectors = nothing)

Perform one time step.

Non-mutating/allocating/out-of-place version.

See also timestep!.

IncompressibleNavierStokes.timestep! — Function

step!(stepper, Δt; cache, momentum_cache, bc_vectors = nothing)

Perform one time step>

Mutating/non-allocating/in-place version.

See also timestep.

Utils

IncompressibleNavierStokes.get_lims — Function

get_lims(x, n = 1.5)

Get approximate lower and upper limits of a field x based on the mean and standard deviation ($\mu \pm n \sigma$). If x is constant, a margin of 1e-4 is enforced. This is required for contour plotting functions that require a certain range.

IncompressibleNavierStokes.plotmat — Function

plotmat(A)

Plot matrix.

Other

IncompressibleNavierStokes.DirichletBC — Type

DirichletBC()

No split boundary conditions, where all velocity components are zero.

DirichletBC(u, dudt)

Dirichlet boundary conditions for the velocity, where u[1] = (x..., t) -> u1_BC up to u[d] = (x..., t) -> ud_BC, where d is the dimension.

To make the pressure the same order as velocity, also provide dudt.

IncompressibleNavierStokes.SymmetricBC — Type

SymmetricBC()

Symmetric boundary conditions. The parallel velocity and pressure is the same at each side of the boundary. The normal velocity is zero.

IncompressibleNavierStokes.PressureBC — Type

PressureBC()

Pressure boundary conditions. The pressure is prescribed on the boundary (usually an "outlet"). The velocity has zero Neumann conditions.

Note: Currently, the pressure is prescribed with the constant value of zero on the entire boundary.

IncompressibleNavierStokes.kinetic_energy — Function

kinetic_energy(setup, u)

Compute total kinetic energy. The velocity components are interpolated to the volume centers and squared.

IncompressibleNavierStokes.FourierLayer — Type

FourierLayer(dimension, kmax, cin => cout; σ = identity, init_weight = glorot_uniform)

Fourier layer operating on uniformly discretized functions.

Some important sizes:

dimension: Spatial dimension, e.g. Dimension(2) or Dimension(3).

(nx..., cin, nsample): Input size

(nx..., cout, nsample): Output size

nx = fill(n, dimension()): Number of points in each spatial dimension

n ≥ kmax: Same number of points in each spatial dimension, must be larger than cut-off wavenumber

kmax: Cut-off wavenumber

nsample: Number of input samples (treated independently)

IncompressibleNavierStokes.offset_p — Function

offset_p(bc)

Number of non-DOF pressure components at boundary.

IncompressibleNavierStokes.momentum_allstage — Function

momentum_allstage(Vⱼ, ϕⱼ, pⱼ, tⱼ, setup; bc_vectors, nstage, get_jacobian = false)

Call momentum for multiple (Vⱼ, pⱼ) pairs, as required in implicit RK methods.

Non-mutating/allocating/out-of-place version.

See also momentum_allstage!.

IncompressibleNavierStokes.momentum_allstage! — Function

momentum_allstage!(
    Fⱼ,
    ∇Fⱼ,
    Vⱼ,
    ϕⱼ,
    pⱼ,
    tⱼ,
    setup,
    cache,
    momentum_cache;
    bc_vectors,
    nstage,
    get_jacobian = false,
)

Call momentum for multiple (V, p) pairs, as required in implicit RK methods.

Mutating/non-allocating/in-place version.

See also momentum_allstage.

IncompressibleNavierStokes.offset_u — Function

offset_u(bc, isnormal, atend)

Number of non-DOF velocity components at boundary. If isnormal, then the velocity is normal to the boundary, else parallel. If atend, it is at the end/right/rear/top boundary, otherwise beginning.

IncompressibleNavierStokes.pressuregradient — Function

pressuregradient(p, setup)

Compute pressure gradient.

IncompressibleNavierStokes.pressuregradient! — Function

pressuregradient!(G, p, setup)

Compute pressure gradient (in-place).

IncompressibleNavierStokes.momentum — Function

momentum(u, t, setup)

Right hand side of momentum equations, excluding pressure gradient.

IncompressibleNavierStokes.Dfield! — Function

Dfield!(d, G, p, setup; ϵ = eps(eltype(p)))

Compute the $D$-field [7] given by

\[D = \frac{2 | \nabla p |}{\nabla^2 p}.\]

IncompressibleNavierStokes.Qfield! — Function

Qfield!(Q, u, setup; ϵ = eps(eltype(Q)))

Compute $Q$-field [8] given by

\[Q = - \frac{1}{2} \sum_{α, β} \frac{\partial u^α}{\partial x^β} \frac{\partial u^β}{\partial x^α}.\]

IncompressibleNavierStokes.Offset — Type

δ = Offset{D}()

Carsesian index unit vector in D = 2 or D = 3 dimensions. Calling δ(α) returns a Cartesian index with 1 in the dimension α and zeros elsewhere.

See https://b-fg.github.io/2023/05/07/waterlily-on-gpu.html for writing kernel loops using Cartesian indices.