API Reference

IncompressibleNavierStokes.IncompressibleNavierStokes — Module

IncompressibleNavierStokes

Energy-conserving solvers for the incompressible Navier-Stokes equations.

IncompressibleNavierStokes.Setup — Function

Setup(
    x, y;
    Re = 1000,
    viscosity_model = LaminarModel(),
    convection_model = NoRegConvectionModel(),
    u_bc = (x, y, t) -> 0,
    v_bc = (x, y, t) -> 0,
    dudt_bc = nothing,
    dvdt_bc = nothing,
    bc_type = (;
        u = (; x = (:periodic, :periodic), y = (:periodic, :periodic)),
        v = (; x = (:periodic, :periodic), y = (:periodic, :periodic)),
    ),
    order4 = false,
    bodyforce_u = (x, y) -> 0,
    bodyforce_v = (x, y) -> 0,
    closure_model = V -> zero(V),
)

Create 2D setup.

Setup(
    x, y, z;
    Re = convert(eltype(x), 1000),
    viscosity_model = LaminarModel(),
    convection_model = NoRegConvectionModel(),
    u_bc = (x, y, w, t) -> 0.0,
    v_bc = (x, y, w, t) -> 0.0,
    w_bc = (x, y, w, t) -> 0.0,
    dudt_bc = nothing,
    dvdt_bc = nothing,
    dwdt_bc = nothing,
    bc_type = (;
        u = (;
            x = (:periodic, :periodic),
            y = (:periodic, :periodic),
            z = (:periodic, :periodic),
        ),
        v = (;
            x = (:periodic, :periodic),
            y = (:periodic, :periodic),
            z = (:periodic, :periodic),
        ),
        w = (;
            x = (:periodic, :periodic),
            y = (:periodic, :periodic),
            z = (:periodic, :periodic),
        ),
        ν = (;
            x = (:periodic, :periodic),
            y = (:periodic, :periodic),
            z = (:periodic, :periodic),
        ),
    ),
    order4 = false,
    bodyforce_u = (x, y, z) -> 0,
    bodyforce_v = (x, y, z) -> 0,
    bodyforce_w = (x, y, z) -> 0,
    closure_model = V -> zero(V),
)

Create 3D setup.

Boundary conditions

IncompressibleNavierStokes.BoundaryConditions — Function

BoundaryConditions(u_bc, v_bc; T = Float64, bc_unsteady, bc_type, kwargs...)

Create discrete boundary condtions.

Values should either be scalars or vectors. All values (u, v, p, k, e) are defined at (x, y) locations, i.e. the corners of pressure volumes, so they cover the entire domain, including corners.

BoundaryConditions(u_bc, v_bc, w_bc; T = Float64, bc_unsteady, bc_type, kwargs...)

Create discrete boundary condtions.

Values should either be scalars or vectors. All values (u, v, p, k, e) are defined at (x, y, z) locations, i.e. the corners of pressure volumes, so they cover the entire domain, including corners.

IncompressibleNavierStokes.get_bc_vectors — Function

get_bc_vectors(setup, t)

Get boundary condition vectors.

Force

IncompressibleNavierStokes.SteadyBodyForce — Function

SteadyBodyForce(fu, fv, grid)

Two-dimensional steady body force f(x, y) = [fu(x, y), fv(x, y)].

SteadyBodyForce(fu, fv, fw, grid)

Three-dimensional steady body force f(x, y, z) = [fu(x, y, z), fv(x, y, z), fw(x, y, z)].

Grid

IncompressibleNavierStokes.Dimension — Type

Dimension(N)

Represent an N-dimensional space. Returns N when called.

julia> d = Dimension(3)
Dimension{3}()

julia> d()
3

IncompressibleNavierStokes.Grid — Function

Grid(x, y; boundary_conditions, order4 = false)

Create nonuniform Cartesian box mesh x × y with boundary conditions boundary_conditions. If order4 is true, a fourth order mesh is created.

Grid(x, y, z; boundary_conditions, order4 = false)

Create nonuniform Cartesian box mesh x × y × z with boundary conditions boundary_conditions. If order4 is true, a fourth order mesh is created.

IncompressibleNavierStokes.cosine_grid — Function

cosine_grid(a, b, N)

Create a nonuniform grid of N + 1 points from a to b using a cosine profile, i.e.

\[x_i = a + \frac{1}{2} \left( 1 - \cos \left( \pi \frac{i}{n} \right) \right) (b - a), \quad i = 0, \dots, N\]

See also stretched_grid.

IncompressibleNavierStokes.max_size — Function

max_size(grid)

Get size of the largest grid element.

IncompressibleNavierStokes.stretched_grid — Function

stretched_grid(a, b, N, s = 1)

Create a nonuniform grid of N + 1 points from a to b with a stretch factor of s. If s = 1, return a uniform spacing from a to b. Otherwise, return a vector $x \in \mathbb{R}^{N + 1}$ such that $x_n = a + \sum_{i = 1}^n s^{i - 1} h$ for $n = 0, \dots , N$. Setting $x_N = b$ then gives $h = (b - a) \frac{1 - s}{1 - s^N}$, resulting in

\[x_n = a + (b - a) \frac{1 - s^n}{1 - s^N}, \quad n = 0, \dots, N.\]

Note that stretched_grid(a, b, N, s)[n] corresponds to $x_{n - 1}$.

See also cosine_grid.

Visocosity Models

IncompressibleNavierStokes.AbstractViscosityModel — Type

AbstractViscosityModel

Abstract viscosity model.

IncompressibleNavierStokes.LaminarModel — Type

LaminarModel()

Laminar model.

IncompressibleNavierStokes.MixingLengthModel — Type

MixingLengthModel()

Mixing-length model with mixing length lm.

IncompressibleNavierStokes.SmagorinskyModel — Type

SmagorinskyModel(Re, C_s = 0.17)

Smagorinsky-Lilly model with constant C_s.

IncompressibleNavierStokes.QRModel — Type

QR(Re)

QR-model.

Convection Models

IncompressibleNavierStokes.AbstractConvectionModel — Type

AbstractConvectionModel

Abstract convection model.

IncompressibleNavierStokes.NoRegConvectionModel — Type

NoRegConvection()

Unregularized convection model.

IncompressibleNavierStokes.C2ConvectionModel — Type

C2ConvectionModel()

C2 regularization convection model.

IncompressibleNavierStokes.C4ConvectionModel — Type

C4ConvectionModel()

C4 regularization convection model.

IncompressibleNavierStokes.LerayConvectionModel — Type

LerayConvectionModel()

Leray regularization convection model.

Momentum

IncompressibleNavierStokes.MomentumCache — Function

MomentumCache(setup)

Preallocation structure for terms in the momentum equations.

IncompressibleNavierStokes.check_symmetry — Function

check_symmetry(V, t, setup, ϵ = 1e-14)

Check symmetry of convection operator.

flag = 0: no symmetry error flag = 1: symmetry error

IncompressibleNavierStokes.compute_conservation — Function

compute_conservation(V, t, setup; bc_vectors = nothing)

Compute mass, momentum and energy conservation properties of velocity field.

IncompressibleNavierStokes.convection — Function

convection(
    model, V, ϕ, setup;
    bc_vectors,
    get_jacobian = false,
    newton_factor = false,
)

Evaluate convective terms c and, optionally, Jacobian ∇c = ∂c/∂V, using the convection model model. The convected quantity is ϕ (usually ϕ = V).

Non-mutating/allocating/out-of-place version.

See also convection!.

IncompressibleNavierStokes.convection! — Function

convection!(
    model, c, ∇c, V, ϕ, setup, cache;
    bc_vectors,
    get_jacobian = false,
    newton_factor = false,
)

Evaluate convective terms c and, optionally, Jacobian ∇c = ∂c/∂V, using the convection model model. The convected quantity is ϕ (usually ϕ = V).

Mutating/non-allocating/in-place version.

See also convection.

IncompressibleNavierStokes.convection_components — Function

convection_components(
    V, ϕ, setup;
    bc_vectors,
    get_jacobian = false,
    newton_factor = false,
    order4 = false,
)

Compute convection components.

Non-mutating/allocating/out-of-place version.

See also convection_components!.

IncompressibleNavierStokes.convection_components! — Function

convection_components!(
    c, ∇c, V, ϕ, setup, cache;
    bc_vectors,
    get_jacobian = false,
    newton_factor = false,
    order4 = false,
)

Compute convection components.

Mutating/non-allocating/in-place version.

See also convection_components.

IncompressibleNavierStokes.diffusion — Function

diffusion!(model, V, setup; bc_vectors, get_jacobian = false)

Evaluate diffusive terms d and optionally Jacobian ∇d = ∂d/∂V using viscosity model model.

Non-mutating/allocating/out-of-place version.

See also diffusion!.

IncompressibleNavierStokes.diffusion! — Function

diffusion!(model, d, ∇d, V, setup; bc_vectors, get_jacobian = false)

Evaluate diffusive terms d and optionally Jacobian ∇d = ∂d/∂V using viscosity model model.

IncompressibleNavierStokes.momentum — Function

momentum(
    V, ϕ, p, t, setup;
    bc_vectors = nothing,
    get_jacobian = false,
    nopressure = false,
    newton_factor = false,
)

Calculate RHS of momentum equations and, optionally, Jacobian with respect to velocity field.

V: velocity field

ϕ: convected field: e.g. $\frac{\partial (\phi_x V)}{\partial x} + \frac{\partial (\phi_y V)}{\partial y}$; usually ϕ = V (so ϕx = u, ϕy = v)

p: pressure

bc_vectors: boundary condition vectors y

get_jacobian: return ∇F = ∂F/∂V

nopressure: exclude pressure gradient; in this case input argument p is not used

newton_factor

Non-mutating/allocating/out-of-place version.

See also momentum!.

IncompressibleNavierStokes.momentum! — Function

momentum!(F, ∇F, V, ϕ, p, t, setup, cache; get_jacobian = false, nopressure = false)

Calculate rhs of momentum equations and, optionally, Jacobian with respect to velocity field.

V: velocity field

ϕ: convected field: e.g. $\frac{\partial (\phi_x V)}{\partial x} + \frac{\partial (\phi_y V)}{\partial y}$; usually ϕ = V (so ϕx = u, ϕy = v)

p: pressure

bc_vectors: boundary condition vectors y

get_jacobian: return ∇F = ∂F/∂V

nopressure: exclude pressure gradient; in this case input argument p is not used

newton_factor

Mutating/non-allocating/in-place version.

See also momentum.

IncompressibleNavierStokes.momentum_allstage — Function

momentum_allstage(Vⱼ, ϕⱼ, pⱼ, tⱼ, setup; bc_vectors, nstage, get_jacobian = false)

Call momentum for multiple (Vⱼ, pⱼ) pairs, as required in implicit RK methods.

Non-mutating/allocating/out-of-place version.

See also momentum_allstage!.

IncompressibleNavierStokes.momentum_allstage! — Function

momentum_allstage!(
    Fⱼ,
    ∇Fⱼ,
    Vⱼ,
    ϕⱼ,
    pⱼ,
    tⱼ,
    setup,
    cache,
    momentum_cache;
    bc_vectors,
    nstage,
    get_jacobian = false,
)

Call momentum for multiple (V, p) pairs, as required in implicit RK methods.

Mutating/non-allocating/in-place version.

See also momentum_allstage.

IncompressibleNavierStokes.strain_tensor — Function

strain_tensor(V, setup; bc_vectors, get_jacobian = false, get_S_abs = false)

Evaluate rate of strain tensor S(V) and its magnitude.

IncompressibleNavierStokes.turbulent_K — Function

turbulent_K(model, setup)

Compute the constant part of the turbulent viscosity.

IncompressibleNavierStokes.turbulent_viscosity — Function

turbulent_viscosity(model, setup, S_abs)

Compute turbulent viscosity based on S_abs.

Operators

IncompressibleNavierStokes.Operators — Function

Operators(grid, boundary_conditions, viscosity_model)

Build operators.

IncompressibleNavierStokes.operator_averaging — Function

operator_averaging(grid, boundary_conditions)

Construct averaging operators.

IncompressibleNavierStokes.operator_convection_diffusion — Function

operator_convection_diffusion(grid, boundary_conditions)

Construct convection and diffusion operators.

IncompressibleNavierStokes.operator_divergence — Function

operator_divergence(grid, boundary_conditions)

Construct divergence and gradient operator.

IncompressibleNavierStokes.operator_interpolation — Function

operator_interpolation(grid, boundary_conditions)

Construct interpolation operators.

IncompressibleNavierStokes.operator_postprocessing — Function

operator_postprocessing(grid, boundary_conditions)

Construct postprocessing operators such as vorticity.

IncompressibleNavierStokes.operator_regularization — Function

operator_regularization(grid, operators)

Build regularization matrices.

IncompressibleNavierStokes.operator_turbulent_diffusion — Function

operator_turbulent_diffusion(grid, boundary_conditions)

Average (turbulent) viscosity to cell faces: from ν at xp, yp to ν at ux, uy, vx, vy locations.

IncompressibleNavierStokes.operator_viscosity — Function

operator_viscosity(viscosity_model, grid, boundary_conditions)

Classical turbulence modelling via the diffusive term

IncompressibleNavierStokes.operator_filter — Function

operator_filter(grid, boundary_conditions)

Construct filtering operator.

Postprocess

IncompressibleNavierStokes.get_streamfunction — Function

get_streamfunction(setup, V, t)

Compute streamfunction $\psi$ from a Poisson equation $\nabla^2 \psi = -\omega$.

IncompressibleNavierStokes.get_velocity — Function

get_velocity(setup, V, t)

Get velocity values at pressure points. Interpolate velocities to pressure positions using BMx and BMy (and BMz), constructed in operator_divergence.jl.

IncompressibleNavierStokes.get_vorticity — Function

get_vorticity(setup, V, t)

Get vorticity from velocity field.

IncompressibleNavierStokes.vorticity! — Function

vorticity!(ω, V, t, setup)

Compute vorticity values at pressure midpoints. This should be consistent with operator_postprocessing.jl.

IncompressibleNavierStokes.plot_force — Function

plot_force(setup, t; kwargs...)

Plot body force.

IncompressibleNavierStokes.plot_grid — Function

plot_grid(x, y)
plot_grid(x, y, z)
plot_grid(grid)

Plot nonuniform Cartesian grid.

IncompressibleNavierStokes.plot_pressure — Function

plot_pressure(setup, p; kwargs...)

Plot pressure.

IncompressibleNavierStokes.plot_streamfunction — Function

plot_streamfunction(setup, V, t; kwargs...)

Plot streamfunction.

IncompressibleNavierStokes.plot_velocity — Function

plot_velocity(setup, V, t; kwargs...)

Plot velocity.

IncompressibleNavierStokes.plot_vorticity — Function

plot_vorticity(setup, V, t; kwargs...)

Plot vorticity field.

IncompressibleNavierStokes.save_vtk — Function

save_vtk(setup, V, p, t, filename = "output/solution")

Save velocity and pressure field to a VTK file.

In the case of a 2D setup, the velocity field is saved as a 3D vector with a z-component of zero, as this seems to be preferred by ParaView.

Preprocess

IncompressibleNavierStokes.create_initial_conditions — Function

create_initial_conditions(
    setup,
    initial_velocity_u,
    initial_velocity_v,
    [initial_velocity_w,]
    t;
    initial_pressure = nothing,
    pressure_solver = DirectPressureSolver(setup),
)

Create initial vectors at starting time t. If p_initial is a function instead of nothing, calculate compatible IC for the pressure.

IncompressibleNavierStokes.random_field — Function

random_field(
    setup, K;
    A = 1_000_000,
    σ = 30,
    s = 5,
    pressure_solver = DirectPressureSolver(setup),
)

Create random field.

K: Maximum wavenumber

A: Eddy amplitude

σ: Variance

s Wavenumber offset before energy starts decaying

Processors

IncompressibleNavierStokes.step_logger — Function

step_logger(; nupdate = 1)

Create processor that logs time step information.

IncompressibleNavierStokes.vtk_writer — Function

vtk_writer(
    setup;
    nupdate = 1,
    dir = "output",
    filename = "solution",
    fields = (:velocity, :pressure, :vorticity),
)

Create processor that writes the solution every nupdate time steps to a VTK file. The resulting Paraview data collection file is stored in "$dir/$filename.pvd".

IncompressibleNavierStokes.field_saver — Function

field_saver(setup; nupdate = 1)

Create processor that stores the solution every nupdate time step to the vector of vectors V and p. The solution times are stored in the vector t.

IncompressibleNavierStokes.field_plotter — Function

field_plotter(
    setup;
    fieldname = :vorticity,
    type = nothing,
    sleeptime = 0.001,
    alpha = 0.05,
)

Plot the solution every time the state o is updated.

The sleeptime is slept at every update, to give Makie time to update the plot. Set this to nothing to skip sleeping.

Available fieldnames are:

:velocity,

:vorticity,

:streamfunction,

:pressure.

Available plot types for 2D are:

heatmap (default),

contour,

contourf.

Available plot types for 3D are:

contour (default).

The alpha value gets passed to contour in 3D.

IncompressibleNavierStokes.energy_history_plotter — Function

energy_history_plotter(setup)

Create energy history plot, with a history point added every time step_observer is updated.

IncompressibleNavierStokes.energy_spectrum_plotter — Function

energy_spectrum_plotter(setup; nupdate = 1)

Create energy spectrum plot, redrawn every time step_observer is updated.

IncompressibleNavierStokes.animator — Function

animator(setup, path; nupdate = 1, plotter = field_plotter(setup); kwargs...)

Animate a plot of the solution every update iteration. The animation is saved to path, which should have one of the following extensions:

".mkv"

".mp4"

".webm"

".gif"

The plot is determined by a plotter processsor. Addtional kwargs are passed to Makie's VideoStream.

Solvers

IncompressibleNavierStokes.get_timestep — Function

get_timestep(stepper, cfl; bc_vectors)

Estimate time step based on eigenvalues of operators, using Gershgorin.

IncompressibleNavierStokes.solve_unsteady — Function

function solve_unsteady(
    setup, V₀, p₀, tlims;
    method = RK44(; T = eltype(V₀)),
    pressure_solver = DirectPressureSolver(setup),
    Δt = nothing,
    cfl = 1,
    n_adapt_Δt = 1,
    inplace = false,
    processors = (),
    device = identity,
)

Solve unsteady problem using method.

If Δt is a real number, it is rounded such that (t_end - t_start) / Δt is an integer. If Δt = nothing, the time step is chosen every n_adapt_Δt iteration with CFL-number cfl .

Each processor is called after every processor.nupdate time step.

All arrays and operators are passed through the device function. This allows for performing computations on a different device than the host (CPU). To compute on an Nvidia GPU using CUDA, change

solve_unsteady(setup, V₀, p₀, tlims; kwargs...)

to the following:

using CUDA
solve_unsteady(
    setup, V₀, p₀, tlims;
    device = cu,
    kwargs...
)

Note that the state observable passed to the processor.initialize function contains vector living on the device, and you may have to move them back to the host using Array(V) and Array(p) in the processor.

IncompressibleNavierStokes.solve_steady_state — Function

function solve_steady_state(
    setup, V₀, p₀;
    jacobian_type = :newton,
    npicard = 2,
    abstol = 1e-10,
    maxiter = 10,
)

Solve steady state problem of the Navier-Stokes equations. This saddlepoint system arises from linearization of the convective terms.

Each processor is called after every processor.nupdate iteration.

Pressure solvers

IncompressibleNavierStokes.AbstractPressureSolver — Type

AbstractPressureSolver

Pressure solver for the Poisson equation.

IncompressibleNavierStokes.DirectPressureSolver — Type

DirectPressureSolver()

Direct pressure solver using a LU decomposition.

IncompressibleNavierStokes.CGPressureSolver — Type

CGPressureSolver(setup; [abstol], [reltol], [maxiter])

Conjugate gradients iterative pressure solver.

IncompressibleNavierStokes.SpectralPressureSolver — Type

SpectralPressureSolver(setup)

Build spectral pressure solver from setup.

IncompressibleNavierStokes.pressure_additional_solve — Function

pressure_additional_solve(pressure_solver, V, p, t, setup; bc_vectors = nothing)

Do additional pressure solve. This makes the pressure compatible with the velocity field, resulting in same order pressure as velocity.

IncompressibleNavierStokes.pressure_additional_solve! — Function

pressure_additional_solve!(pressure_solver, V, p, t, setup, momentum_cache, F, f; bc_vectors)

Do additional pressure solve. This makes the pressure compatible with the velocity field, resulting in same order pressure as velocity.

IncompressibleNavierStokes.pressure_poisson — Function

pressure_poisson(solver, f)

Solve the Poisson equation for the pressure with right hand side f at time t. For periodic and no-slip BC, the sum of f should be zero.

Non-mutating/allocating/out-of-place version.

See also pressure_poisson!.

IncompressibleNavierStokes.pressure_poisson! — Function

pressure_poisson!(solver, p, f)

Solve the Poisson equation for the pressure with right hand side f at time t. For periodic and no-slip BC, the sum of f should be zero.

Mutating/non-allocating/in-place version.

See also pressure_poisson.

Time steppers

IncompressibleNavierStokes.AbstractODEMethod — Type

AbstractODEMethod

Abstract ODE method.

IncompressibleNavierStokes.AbstractRungeKuttaMethod — Type

AbstractRungeKuttaMethod

Abstract Runge Kutta method.

IncompressibleNavierStokes.AdamsBashforthCrankNicolsonMethod — Type

AdamsBashforthCrankNicolsonMethod(
    T = Float64;
    α₁ = T(3 // 2),
    α₂ = T(-1 // 2),
    θ = T(1 // 2),
    p_add_solve = true,
    method_startup = RK44(; T),
)

IMEX AB-CN: Adams-Bashforth for explicit convection (parameters α₁ and α₂) and Crank-Nicolson for implicit diffusion (implicitness θ). The method is second order for θ = 1/2.

Adams-Bashforth for convection and Crank-Nicolson for diffusion formulation:

\[\begin{align*} (\mathbf{u}^{n+1} - \mathbf{u}^n) / Δt & = -(\alpha_1 \mathbf{c}^n + \alpha_2 \mathbf{c}^{n-1}) \\ & + \theta \mathbf{d}^{n+1} + (1-\theta) \mathbf{d}^n \\ & + \theta \mathbf{F}^{n+1} + (1-\theta) \mathbf{F}^n \\ & + \theta \mathbf{BC}^{n+1} + (1-\theta) \mathbf{BC}^n \\ & - \mathbf{G} \mathbf{p} + \mathbf{y}_p \end{align*}\]

where BC are boundary conditions of diffusion. This is rewritten as:

\[\begin{align*} (\frac{1}{\Delta t} \mathbf{I} - \theta \mathbf{D}) \mathbf{u}^{n+1} & = (\frac{1}{\Delta t} \mathbf{I} - (1 - \theta) \mathbf{D}) \mathbf{u}^{n} \\ & - (\alpha_1 \mathbf{c}^n + \alpha_2 \mathbf{c}^{n-1}) \\ & + \theta \mathbf{F}^{n+1} + (1-\theta) \mathbf{F}^{n} \\ & + \theta \mathbf{BC}^{n+1} + (1-\theta) \mathbf{BC}^{n} \\ & - \mathbf{G} \mathbf{p} + \mathbf{y}_p \end{align*}\]

The LU decomposition of the LHS matrix is computed every time the time step changes.

Note that, in constrast to explicit methods, the pressure from previous time steps has an influence on the accuracy of the velocity.

IncompressibleNavierStokes.OneLegMethod — Type

OneLegMethod(
    T = Float64;
    β = T(1 // 2),
    p_add_solve = true,
    method_startup = RK44(; T),
)

Explicit one-leg β-method following symmetry-preserving discretization of turbulent flow. See [Verstappen and Veldman (JCP 2003)] for details, or [Direct numerical simulation of turbulence at lower costs (Journal of Engineering Mathematics 1997)].

Formulation:

\[\frac{(\beta + 1/2) u^{n+1} - 2 \beta u^{n} + (\beta - 1/2) u^{n-1}}{\Delta t} = F((1 + \beta) u^n - \beta u^{n-1}).\]

IncompressibleNavierStokes.ExplicitRungeKuttaMethod — Type

ExplicitRungeKuttaMethod(; A, b, c, r, p_add_solve = true)

Explicit Runge Kutta method.

IncompressibleNavierStokes.ImplicitRungeKuttaMethod — Type

ImplicitRungeKuttaMethod(;
    A,
    b,
    c,
    r,
    newton_type = :full,
    maxiter = 10,
    abstol = 1e-14,
    reltol = 1e-14,
    p_add_solve = true,
)

Implicit Runge Kutta method.

The implicit linear system is solved at each time step using Newton's method. The newton_type may be one of the following:

:no: Replace iteration matrix with I/Δt (no Jacobian)

:approximate: Build Jacobian once before iterations only

:full: Build Jacobian at each iteration

IncompressibleNavierStokes.isexplicit — Function

isexplicit(method)

Return true if method is explicit, i.e. the value at a certain time step is given explicitly as a function of the previous time steps only.

IncompressibleNavierStokes.lambda_conv_max — Function

lambda_conv_max(method)

Get maximum value of stability region for the convection operator (not a very good indication for the methods that do not include the imaginary axis)

IncompressibleNavierStokes.lambda_diff_max — Function

lambda_diff_max(method)

Get maximum value of stability region for the diffusion operator.

IncompressibleNavierStokes.nstage — Function

nstage(rk_method)

Get number of stages of the Runge-Kutta method.

IncompressibleNavierStokes.ode_method_cache — Function

ode_method_cache(method, setup)

Get time stepper cache for the given ODE method.

IncompressibleNavierStokes.runge_kutta_method — Function

runge_kutta_method(A, b, c, r; [p_add_solve], [newton_type], [maxiter], [abstol], [reltol])

Get Runge Kutta method. The function checks whether the method is explicit.

p_add_solve: whether to add a pressure solve step to the method.

For implicit RK methods: newton_type, maxiter, abstol, reltol.

IncompressibleNavierStokes.step — Function

step(stepper, Δt; bc_vectors = nothing)

Perform one time step.

Non-mutating/allocating/out-of-place version.

See also step!.

IncompressibleNavierStokes.step! — Function

step!(stepper, Δt; cache, momentum_cache, bc_vectors = nothing)

Perform one time step>

Mutating/non-allocating/in-place version.

See also step.

Utils

IncompressibleNavierStokes.filter_convection — Function

filter_convection(u, diff_matrix, bc, α)

Construct filter for convective terms.

Non-mutating/allocating/out-of-place version.

See also filter_convection!.

IncompressibleNavierStokes.filter_convection! — Function

filter_convection!(ū, u, diff_matrix, bc, α)

Construct filter for convective terms.

Mutating/non-allocating/in-place version.

See also filter_convection.

IncompressibleNavierStokes.get_lims — Function

get_lims(x, n = 1.5)

Get approximate lower and upper limits of a field x based on the mean and standard deviation ($\mu \pm n \sigma$). If x is constant, a margin of 1e-4 is enforced. This is required for contour plotting functions that require a certain range.