API Reference

IncompressibleNavierStokes.IncompressibleNavierStokes — Module

IncompressibleNavierStokes

Energy-conserving solvers for the incompressible Navier-Stokes equations.

IncompressibleNavierStokes.Setup — Function

Setup(
    x...;
    boundary_conditions = ntuple(d -> (PeriodicBC(), PeriodicBC()), length(x)),
    Re = convert(eltype(x[1]), 1_000),
    viscosity_model = LaminarModel(),
    convection_model = NoRegConvectionModel(),
    bodyforce = nothing,
    closure_model = nothing,
    ArrayType = Array,
)

Create setup.

Grid

IncompressibleNavierStokes.Dimension — Type

Dimension(N)

Represent an N-dimensional space. Returns N when called.

julia> d = Dimension(3)
Dimension{3}()

julia> d()
3

IncompressibleNavierStokes.Grid — Function

Grid(x, boundary_conditions)

Create nonuniform Cartesian box mesh x[1] × ... × x[d] with boundary conditions boundary_conditions.

IncompressibleNavierStokes.cosine_grid — Function

cosine_grid(a, b, N)

Create a nonuniform grid of N + 1 points from a to b using a cosine profile, i.e.

\[x_i = a + \frac{1}{2} \left( 1 - \cos \left( \pi \frac{i}{n} \right) \right) (b - a), \quad i = 0, \dots, N\]

See also stretched_grid.

IncompressibleNavierStokes.max_size — Function

max_size(grid)

Get size of the largest grid element.

IncompressibleNavierStokes.stretched_grid — Function

stretched_grid(a, b, N, s = 1)

Create a nonuniform grid of N + 1 points from a to b with a stretch factor of s. If s = 1, return a uniform spacing from a to b. Otherwise, return a vector $x \in \mathbb{R}^{N + 1}$ such that $x_n = a + \sum_{i = 1}^n s^{i - 1} h$ for $n = 0, \dots , N$. Setting $x_N = b$ then gives $h = (b - a) \frac{1 - s}{1 - s^N}$, resulting in

\[x_n = a + (b - a) \frac{1 - s^n}{1 - s^N}, \quad n = 0, \dots, N.\]

Note that stretched_grid(a, b, N, s)[n] corresponds to $x_{n - 1}$.

See also cosine_grid.

Convection Models

IncompressibleNavierStokes.AbstractConvectionModel — Type

AbstractConvectionModel

Abstract convection model.

IncompressibleNavierStokes.NoRegConvectionModel — Type

NoRegConvection()

Unregularized convection model.

IncompressibleNavierStokes.C2ConvectionModel — Type

C2ConvectionModel()

C2 regularization convection model.

IncompressibleNavierStokes.C4ConvectionModel — Type

C4ConvectionModel()

C4 regularization convection model.

IncompressibleNavierStokes.LerayConvectionModel — Type

LerayConvectionModel()

Leray regularization convection model.

Postprocess

IncompressibleNavierStokes.plot_force — Function

plot_force(setup, t; kwargs...)

Plot body force.

IncompressibleNavierStokes.plot_grid — Function

plot_grid(x, y)
plot_grid(x, y, z)
plot_grid(grid)

Plot nonuniform Cartesian grid.

IncompressibleNavierStokes.plot_pressure — Function

plot_pressure(setup, p; kwargs...)

Plot pressure.

IncompressibleNavierStokes.plot_streamfunction — Function

plot_streamfunction(setup, V, t; kwargs...)

Plot streamfunction.

IncompressibleNavierStokes.plot_velocity — Function

plot_velocity(setup, V, t; kwargs...)

Plot velocity.

IncompressibleNavierStokes.plot_vorticity — Function

plot_vorticity(setup, V, t; kwargs...)

Plot vorticity field.

IncompressibleNavierStokes.save_vtk — Function

save_vtk(setup, u, p, filename = "output/solution")

Save velocity and pressure field to a VTK file.

In the case of a 2D setup, the velocity field is saved as a 3D vector with a z-component of zero, as this seems to be preferred by ParaView.

Preprocess

IncompressibleNavierStokes.create_initial_conditions — Function

create_initial_conditions(
    setup,
    initial_velocity,
    t = 0;
    pressure_solver = DirectPressureSolver(setup),
)

Create initial vectors (u, p) at starting time t.

IncompressibleNavierStokes.random_field — Function

random_field(
    setup, t;
    A = setup.grid.N[1] * 10_000,
    σ = 30,
    s = 5,
    pressure_solver = DirectPressureSolver(setup),
)

Create random field.

K: Maximum wavenumber

A: Eddy amplitude

σ: Variance

s Wavenumber offset before energy starts decaying

Processors

IncompressibleNavierStokes.step_logger — Function

step_logger(; nupdate = 1)

Create processor that logs time step information.

IncompressibleNavierStokes.vtk_writer — Function

vtk_writer(
    setup;
    nupdate = 1,
    dir = "output",
    filename = "solution",
    fields = (:velocity, :pressure, :vorticity),
)

Create processor that writes the solution every nupdate time steps to a VTK file. The resulting Paraview data collection file is stored in "$dir/$filename.pvd".

IncompressibleNavierStokes.field_saver — Function

field_saver(setup; nupdate = 1)

Create processor that stores the solution every nupdate time step to the vector of vectors V and p. The solution times are stored in the vector t.

IncompressibleNavierStokes.field_plotter — Function

field_plotter(
    setup;
    fieldname = :vorticity,
    type = nothing,
    sleeptime = 0.001,
    alpha = 0.05,
)

Plot the solution every time the state o is updated.

The sleeptime is slept at every update, to give Makie time to update the plot. Set this to nothing to skip sleeping.

Available fieldnames are:

:velocity,

:vorticity,

:streamfunction,

:pressure.

Available plot types for 2D are:

heatmap (default),

image,

contour,

contourf.

Available plot types for 3D are:

contour (default).

The alpha value gets passed to contour in 3D.

IncompressibleNavierStokes.energy_history_plotter — Function

energy_history_plotter(setup)

Create energy history plot, with a history point added every time step_observer is updated.

IncompressibleNavierStokes.energy_spectrum_plotter — Function

energy_spectrum_plotter(setup; nupdate = 1)

Create energy spectrum plot, redrawn every time step_observer is updated.

IncompressibleNavierStokes.animator — Function

animator(setup, path; nupdate = 1, plotter = field_plotter(setup); kwargs...)

Animate a plot of the solution every update iteration. The animation is saved to path, which should have one of the following extensions:

".mkv"

".mp4"

".webm"

".gif"

The plot is determined by a plotter processsor. Addtional kwargs are passed to Makie's VideoStream.

Solvers

IncompressibleNavierStokes.get_timestep — Function

get_timestep(stepper, cfl; bc_vectors)

Estimate time step based on eigenvalues of operators, using Gershgorin.

IncompressibleNavierStokes.solve_unsteady — Function

function solve_unsteady(
    setup, u₀, p₀, tlims;
    method = RK44(; T = eltype(u₀)),
    pressure_solver = DirectPressureSolver(setup),
    Δt = zero(eltype(u₀)),
    cfl = 1,
    n_adapt_Δt = 1,
    inplace = true,
    processors = (),
)

Solve unsteady problem using method.

If Δt is a real number, it is rounded such that (t_end - t_start) / Δt is an integer. If Δt = nothing, the time step is chosen every n_adapt_Δt iteration with CFL-number cfl .

Each processor is called after every processor.nupdate time step.

All arrays and operators are passed through the device function. This allows for performing computations on a different device than the host (CPU). To compute on an Nvidia GPU using CUDA, change

solve_unsteady(setup, u₀, p₀, tlims; kwargs...)

to the following:

using CUDA
solve_unsteady(
    setup, u₀, p₀, tlims;
    device = cu,
    kwargs...
)

Note that the state observable passed to the processor.initialize function contains vector living on the device, and you may have to move them back to the host using Array(u) and Array(p) in the processor.

IncompressibleNavierStokes.solve_steady_state — Function

function solve_steady_state(
    setup, V₀, p₀;
    jacobian_type = :newton,
    npicard = 2,
    abstol = 1e-10,
    maxiter = 10,
)

Solve steady state problem of the Navier-Stokes equations. This saddlepoint system arises from linearization of the convective terms.

Each processor is called after every processor.nupdate iteration.

Utils

IncompressibleNavierStokes.get_lims — Function

get_lims(x, n = 1.5)

Get approximate lower and upper limits of a field x based on the mean and standard deviation ($\mu \pm n \sigma$). If x is constant, a margin of 1e-4 is enforced. This is required for contour plotting functions that require a certain range.

IncompressibleNavierStokes.plotmat — Function

plotmat(A)

Plot matrix.