API Reference

IncompressibleNavierStokes.IncompressibleNavierStokes — Module

IncompressibleNavierStokes

Energy-conserving solvers for the incompressible Navier-Stokes equations.

IncompressibleNavierStokes.Setup — Type

Setup(grid, boundary_conditions, viscosity_model, convection_model, force, operators)

Simulation setup.

Boundary conditions

IncompressibleNavierStokes.BoundaryConditions — Type

BoundaryConditions{T}

Boundary conditions with floating point type T.

IncompressibleNavierStokes.get_bc_vectors — Function

get_bc_vectors(setup, t)

Get boundary condition vectors.

Force

IncompressibleNavierStokes.SteadyBodyForce — Type

SteadyBodyForce{F1,F2,F3,T}

Steady (constant) body force.

IncompressibleNavierStokes.UnsteadyBodyForce — Type

UnsteadyBodyForce{F1,F2,F3,T}

Unsteady (time-dependent) body force.

Grid

IncompressibleNavierStokes.Grid — Type

Grid{T,N}()

Nonuniform Cartesian grid of dimension N and floating point type T.

IncompressibleNavierStokes.cosine_grid — Function

cosine_grid(a, b, N)

Create a nonuniform grid of N + 1 points from a to b using a cosine profile, i.e.

\[x_i = a + \frac{1}{2} \left( 1 - \cos \left( \pi \frac{i}{n} \right) \right) (b - a), \quad i = 0, \dots, N\]

See also stretched_grid.

IncompressibleNavierStokes.get_dimension — Function

get_dimension(grid)

Get dimension of grid.

IncompressibleNavierStokes.max_size — Function

max_size(grid)

Get size of the largest grid element.

IncompressibleNavierStokes.stretched_grid — Function

stretched_grid(a, b, N, s = 1)

Create a nonuniform grid of N + 1 points from a to b with a stretch factor of s. If s = 1, return a uniform spacing from a to b. Otherwise, return a vector $x \in \mathbb{R}^{N + 1}$ such that $x_n = a + \sum_{i = 1}^n s^{i - 1} h$ for $n = 0, \dots , N$. Setting $x_N = b$ then gives $h = (b - a) \frac{1 - s}{1 - s^N}$, resulting in

\[x_n = a + (b - a) \frac{1 - s^n}{1 - s^N}, \quad n = 0, \dots, N.\]

Note that stretched_grid(a, b, N, s)[n] corresponds to $x_{n - 1}$.

See also cosine_grid.

Visocosity Models

IncompressibleNavierStokes.AbstractViscosityModel — Type

AbstractViscosityModel

Abstract viscosity model.

IncompressibleNavierStokes.LaminarModel — Type

LaminarModel(Re)

Laminar model with Reynolds number Re.

IncompressibleNavierStokes.KEpsilonModel — Type

KEpsilonModel(Re)

K-epsilon model with Reynolds number Re.

IncompressibleNavierStokes.MixingLengthModel — Type

MixingLengthModel(Re)

Mixing-length model with Reynolds number Re and mixing length lm.

IncompressibleNavierStokes.SmagorinskyModel — Type

SmagorinskyModel(Re, C_s = 0.17)

Smagorinsky-Lilly model with Reynolds number Re and constant C_s.

IncompressibleNavierStokes.QRModel — Type

QR(Re)

QR-model with Reynolds number Re.

Convection Models

IncompressibleNavierStokes.AbstractConvectionModel — Type

AbstractConvectionModel

Abstract convection model.

IncompressibleNavierStokes.NoRegConvectionModel — Type

NoRegConvection()

Unregularized convection model.

IncompressibleNavierStokes.C2ConvectionModel — Type

C2ConvectionModel()

C2 regularization convection model.

IncompressibleNavierStokes.C4ConvectionModel — Type

C4ConvectionModel()

C4 regularization convection model.

IncompressibleNavierStokes.LerayConvectionModel — Type

LerayConvectionModel()

Leray regularization convection model.

Momentum

IncompressibleNavierStokes.MomentumCache — Type

MomentumCache

Preallocation structure for terms in the momentum equations.

IncompressibleNavierStokes.bodyforce — Function

bodyforce(force, V, t, setup; get_jacobian = false)

Compute body force F in momentum equations at velocity points.

Non-mutating/allocating/out-of-place version.

See also bodyforce!.

IncompressibleNavierStokes.bodyforce! — Function

bodyforce!(force, F, t, setup)

Compute body force F in momentum equations at velocity points.

Mutating/non-allocating/in-place version.

See also bodyforce.

IncompressibleNavierStokes.check_symmetry — Function

check_symmetry(V, t, setup, ϵ = 1e-14)

Check symmetry of convection operator.

flag = 0: no symmetry error flag = 1: symmetry error

IncompressibleNavierStokes.compute_conservation — Function

compute_conservation(V, t, setup; bc_vectors = nothing)

Compute mass, momentum and energy conservation properties of velocity field.

IncompressibleNavierStokes.convection — Function

convection(model, V, ϕ, setup; bc_vectors, get_jacobian = false)

Evaluate convective terms c and, optionally, Jacobian ∇c = ∂c/∂V, using the convection model model. The convected quantity is ϕ (usually ϕ = V).

Non-mutating/allocating/out-of-place version.

See also convection!.

IncompressibleNavierStokes.convection! — Function

convection!(model, c, ∇c, V, ϕ, setup, cache; bc_vectors, get_jacobian = false)

Evaluate convective terms c and, optionally, Jacobian ∇c = ∂c/∂V, using the convection model model. The convected quantity is ϕ (usually ϕ = V).

Mutating/non-allocating/in-place version.

See also convection.

IncompressibleNavierStokes.convection_components — Function

convection_components(
    V, ϕ, setup;
    bc_vectors,
    get_jacobian = false,
    newton_factor = false,
    order4 = false,
)

Compute convection components.

Non-mutating/allocating/out-of-place version.

See also convection_components!.

IncompressibleNavierStokes.convection_components! — Function

convection_components!(
    c, ∇c, V, ϕ, setup, cache;
    bc_vectors,
    get_jacobian = false,
    newton_factor = false,
    order4 = false,
)

Compute convection components.

Mutating/non-allocating/in-place version.

See also convection_components.

IncompressibleNavierStokes.diffusion — Function

diffusion!(model, V, setup; bc_vectors, get_jacobian = false)

Evaluate diffusive terms d and optionally Jacobian ∇d = ∂d/∂V using viscosity model model.

Non-mutating/allocating/out-of-place version.

See also diffusion!.

IncompressibleNavierStokes.diffusion! — Function

diffusion!(model, d, ∇d, V, setup; bc_vectors, get_jacobian = false)

Evaluate diffusive terms d and optionally Jacobian ∇d = ∂d/∂V using viscosity model model.

IncompressibleNavierStokes.momentum — Function

momentum(
    V, ϕ, p, t, setup;
    bc_vectors = nothing,
    get_jacobian = false,
    nopressure = false,
    newton_factor = false,
)

Calculate RHS of momentum equations and, optionally, Jacobian with respect to velocity field.

V: velocity field

ϕ: convected field: e.g. $\frac{\partial (\phi_x V)}{\partial x} + \frac{\partial (\phi_y V)}{\partial y}$; usually ϕ = V (so ϕx = u, ϕy = v)

p: pressure

bc_vectors: boundary condition vectors y

get_jacobian: return ∇F = ∂F/∂V

nopressure: exclude pressure gradient; in this case input argument p is not used

newton_factor

Non-mutating/allocating/out-of-place version.

See also momentum!.

IncompressibleNavierStokes.momentum! — Function

momentum!(F, ∇F, V, ϕ, p, t, setup, cache; get_jacobian = false, nopressure = false)

Calculate rhs of momentum equations and, optionally, Jacobian with respect to velocity field.

V: velocity field

ϕ: convected field: e.g. $\frac{\partial (\phi_x V)}{\partial x} + \frac{\partial (\phi_y V)}{\partial y}$; usually ϕ = V (so ϕx = u, ϕy = v)

p: pressure

bc_vectors: boundary condition vectors y

get_jacobian: return ∇F = ∂F/∂V

nopressure: exclude pressure gradient; in this case input argument p is not used

newton_factor

Mutating/non-allocating/in-place version.

See also momentum.

IncompressibleNavierStokes.momentum_allstage — Function

momentum_allstage(V, C, p, t, setup; get_jacobian = false)

Call momentum for multiple (V, p) pairs, as required in implicit RK methods.

Non-mutating/allocating/out-of-place version.

See also momentum_allstage!.

IncompressibleNavierStokes.momentum_allstage! — Function

momentum_allstage!(F, ∇F, V, C, p, t, setup, cache, momentum_cache; get_jacobian = false)

Call momentum for multiple (V, p) pairs, as required in implicit RK methods.

Mutating/non-allocating/in-place version.

See also momentum_allstage.

IncompressibleNavierStokes.strain_tensor — Function

strain_tensor(V, setup; bc_vectors, get_jacobian = false, get_S_abs = false)

Evaluate rate of strain tensor S(V) and its magnitude.

IncompressibleNavierStokes.turbulent_K — Function

turbulent_K(model, setup)

Compute the constant part of the turbulent viscosity.

IncompressibleNavierStokes.turbulent_viscosity — Function

turbulent_viscosity(model, setup, S_abs)

Compute turbulent viscosity based on S_abs.

Operators

IncompressibleNavierStokes.Operators — Type

Operators()

Discrete operators.

IncompressibleNavierStokes.ke_convection — Function

ke_convection(grid, boundary_conditions)

K-epsilon convection.

IncompressibleNavierStokes.ke_diffusion — Function

ke_diffusion(grid, boundary_conditions)

K-epsilon diffusion.

IncompressibleNavierStokes.ke_production — Function

ke_production(grid, boundary_conditions)

Differencing velocity to k-points.

IncompressibleNavierStokes.ke_viscosity — Function

ke_viscosity(grid, boundary_conditions)

Average (turbulent) viscosity to cell faces.

IncompressibleNavierStokes.operator_averaging — Function

operator_averaging(grid, boundary_conditions)

Construct averaging operators.

IncompressibleNavierStokes.operator_convection_diffusion — Function

operator_convection_diffusion(grid, boundary_conditions, viscosity_model)

Construct convection and diffusion operators.

IncompressibleNavierStokes.operator_divergence — Function

operator_divergence(grid, boundary_conditions)

Construct divergence and gradient operator.

IncompressibleNavierStokes.operator_interpolation — Function

operator_interpolation(grid, boundary_conditions)

Construct averaging operators.

IncompressibleNavierStokes.operator_postprocessing — Function

operator_postprocessing(grid, boundary_conditions)

Construct postprocessing operators such as vorticity.

IncompressibleNavierStokes.operator_regularization — Function

operator_regularization(grid)

Build regularization matrices.

IncompressibleNavierStokes.operator_turbulent_diffusion — Function

operator_turbulent_diffusion(grid, boundary_conditions)

Average (turbulent) viscosity to cell faces: from ν at xp, yp to ν at ux, uy, vx, vy locations.

See also ke_viscosity.jl.

IncompressibleNavierStokes.operator_viscosity — Function

operator_viscosity(model, grid, boundary_conditions)

Classical turbulence modelling via the diffusive term

Postprocess

IncompressibleNavierStokes.get_streamfunction — Function

get_streamfunction(V, t, setup)

Compute streamfunction $\psi$ from a Poisson equation $\nabla^2 \psi = -\omega$.

IncompressibleNavierStokes.get_velocity — Function

get_velocity(V, t, setup)

Get velocity values at pressure points. Interpolate velocities to pressure positions using BMx and BMy (and BMz), constructed in operator_divergence.jl.

IncompressibleNavierStokes.get_vorticity — Function

get_vorticity(V, t, setup)

Get vorticity from velocity field.

IncompressibleNavierStokes.vorticity! — Function

vorticity!(ω, V, t, setup)

Compute vorticity values at pressure midpoints. This should be consistent with operator_postprocessing.jl.

IncompressibleNavierStokes.plot_force — Function

plot_force(setup, t; kwargs...)

Plot body force.

IncompressibleNavierStokes.plot_grid — Function

plot_grid(x, y)
plot_grid(x, y, z)
plot_grid(grid)

Plot nonuniform Cartesian grid.

IncompressibleNavierStokes.plot_pressure — Function

plot_pressure(setup, p; kwargs...)

Plot pressure.

IncompressibleNavierStokes.plot_streamfunction — Function

plot_streamfunction(setup, V, t; kwargs...)

Plot streamfunction.

IncompressibleNavierStokes.plot_tracers — Function

plot_tracers(tracer)

Plot tracer.

IncompressibleNavierStokes.plot_velocity — Function

plot_velocity(setup, V, t; kwargs...)

Plot velocity.

IncompressibleNavierStokes.plot_vorticity — Function

plot_vorticity(setup, V, t; kwargs...)

Plot vorticity field.

IncompressibleNavierStokes.save_vtk — Function

save_vtk(V, p, t, setup, filename = "output/solution")

Save velocity and pressure field to a VTK file.

In the case of a 2D setup, the velocity field is saved as a 3D vector with a z-component of zero, as this seems to be preferred by ParaView.

Preprocess

IncompressibleNavierStokes.create_initial_conditions — Function

create_initial_conditions(
    setup,
    t;
    initial_velocity_u,
    initial_velocity_v,
    [initial_velocity_w,]
    initial_pressure = nothing,
    pressure_solver = DirectPressureSolver(setup),
)

Create initial vectors at starting time t. If p_initial is a function instead of nothing, calculate compatible IC for the pressure.

Problems

IncompressibleNavierStokes.SteadyStateProblem — Type

SteadyStateProblem(setup, V₀, p₀)

Steady state problem. The solution V and p should be such that $\frac{\partial \mathbf{V}}{\partial t} = \mathbf{0}$.

IncompressibleNavierStokes.UnsteadyProblem — Type

UnsteadyProblem(setup, V₀, p₀, (t_start, t_stop))

Unsteady problem with initial conditions V₀, p₀ to be solved from t_start to t_stop.

IncompressibleNavierStokes.is_steady — Function

is_steady(problem)

Returns true if problem is unsteady.

Processors

IncompressibleNavierStokes.AbstractProcessor — Type

AbstractProcessor

Abstract iteration processor.

IncompressibleNavierStokes.Logger — Type

Logger(nupdate)

Print time stepping information after every time step.

IncompressibleNavierStokes.RealTimePlotter — Type

RealTimePlotter(; nupdate, fieldname, type = heatmap)

Plot the solution every nupdate time steps. Available fieldnames are:

:velocity,

:vorticity,

:streamfunction,

:pressure.

Available plot types are:

heatmap,

contour,

contourf.

IncompressibleNavierStokes.VTKWriter — Type

VTKWriter(; nupdate, dir = "output", filename = "solution")

Write the solution every nupdate time steps to a VTK file. The resulting Paraview data collection file is stored in "$dir/$filename.pvd".

IncompressibleNavierStokes.QuantityTracer — Type

QuantityTracer(nupdate)

Store scalar quantities (maximum divergence, momentum, kinetic energy) every nupdate time steps.

IncompressibleNavierStokes.initialize! — Function

initialize!(processor, stepper)

Initialize processor.

IncompressibleNavierStokes.process! — Function

process!(processor, stepper)

Process iteration.

IncompressibleNavierStokes.finalize! — Function

finalize!(processor)

Finalize processing after iterations.

Solvers

IncompressibleNavierStokes.get_timestep — Function

get_timestep(stepper, cfl; bc_vectors)

Estimate time step based on eigenvalues of operators, using Gershgorin.

IncompressibleNavierStokes.solve — Function

solve(
    problem::SteadyStateProblem;
    jacobian_type = :newton,
    npicard = 2,
    abstol = 1e-10,
    maxiter = 10,
)

Solve steady state problem of the Navier-Stokes equations. This saddlepoint system arises from linearization of the convective terms.

Each processor is called after every processor.nupdate iteration.

solve(
    problem::UnsteadyProblem,
    method;
    pressure_solver = DirectPressureSolver(problem.setup),
    Δt = nothing,
    cfl = 1,
    n_adapt_Δt = 1,
    nstartup = 1,
    method_startup = nothing,
    inplace = false,
    processors = [],
)

Solve unsteady problem using method.

If Δt is a real number, it is rounded such that (t_end - t_start) / Δt is an integer. If Δt = nothing, the time step is chosen every n_adapt_Δt iteration with CFL-number cfl .

For methods that are not self-starting, nstartup startup iterations are performed with method_startup.

Each processor is called after every processor.nupdate time step.

IncompressibleNavierStokes.solve_animate — Function

solve_animate(
    problem::UnsteadyProblem,
    method;
    pressure_solver = DirectPressureSolver(problem.setup),
    Δt = nothing,
    cfl = 1,
    n_adapt_Δt = 1,
    method_startup = nothing,
    nstartup = 1,
    animator = RealTimePlotter(),
    filename = "vorticity.gif",
    nframe = 200,
    nsubframe = 4,
    framerate = 20,
)

Solve unsteady problem using method.

If Δt is a real number, it is rounded such that (t_end - t_start) / Δt is an integer. If Δt = nothing, the time step is chosen every n_adapt_Δt iteration with CFL-number cfl .

For methods that are not self-starting, nstartup startup iterations are performed with method_startup.

Each processor is called after every processor.nupdate time step.

Pressure solvers

IncompressibleNavierStokes.AbstractPressureSolver — Type

AbstractPressureSolver

Pressure solver for the Poisson equation.

IncompressibleNavierStokes.DirectPressureSolver — Type

DirectPressureSolver()

Direct pressure solver using a LU decomposition.

IncompressibleNavierStokes.CGPressureSolver — Type

CGPressureSolver(abstol, reltol, maxiter)

Conjugate gradients iterative pressure solver.

IncompressibleNavierStokes.FourierPressureSolver — Type

FourierPressureSolver()

Fourier transform pressure solver for periodic domains.

IncompressibleNavierStokes.pressure_additional_solve — Function

pressure_additional_solve(pressure_solver, V, p, t, setup; bc_vectors = nothing)

Do additional pressure solve. This makes the pressure compatible with the velocity field, resulting in same order pressure as velocity.

IncompressibleNavierStokes.pressure_additional_solve! — Function

pressure_additional_solve!(pressure_solver, V, p, t, setup, momentum_cache, F, f; bc_vectors)

Do additional pressure solve. This makes the pressure compatible with the velocity field, resulting in same order pressure as velocity.

IncompressibleNavierStokes.pressure_poisson — Function

pressure_poisson(solver, f, setup)

Solve the Poisson equation for the pressure with right hand side f at time t. For periodic and no-slip BC, the sum of f should be zero.

Non-mutating/allocating/out-of-place version.

See also pressure_poisson!.

IncompressibleNavierStokes.pressure_poisson! — Function

pressure_poisson!(solver, p, f)

Solve the Poisson equation for the pressure with right hand side f at time t. For periodic and no-slip BC, the sum of f should be zero.

Mutating/non-allocating/in-place version.

See also pressure_poisson.

Time steppers

IncompressibleNavierStokes.AbstractODEMethod — Type

AbstractODEMethod

Abstract ODE method.

IncompressibleNavierStokes.AbstractRungeKuttaMethod — Type

AbstractRungeKuttaMethod

Abstract Runge Kutta method.

IncompressibleNavierStokes.AdamsBashforthCrankNicolsonMethod — Type

AdamsBashforthCrankNicolsonMethod(; α₁ = 3 // 2, α₂ = -1 // 2, θ = 1 // 2, p_add_solve = true)

IMEX AB-CN: Adams-Bashforth for explicit convection (parameters α₁ and α₂) and Crank-Nicolson for implicit diffusion (implicitness θ). The method is second order for θ = 1/2.

Adams-Bashforth for convection and Crank-Nicolson for diffusion formulation:

\[\begin{align*} (\mathbf{u}^{n+1} - \mathbf{u}^n) / Δt & = -(\alpha_1 \mathbf{c}^n + \alpha_2 \mathbf{c}^{n-1}) \\ & + \theta \mathbf{d}^{n+1} + (1-\theta) \mathbf{d}^n \\ & + \theta \mathbf{F}^{n+1} + (1-\theta) \mathbf{F}^n \\ & + \theta \mathbf{BC}^{n+1} + (1-\theta) \mathbf{BC}^n \\ & - \mathbf{G} \mathbf{p} + \mathbf{y}_p \end{align*}\]

where BC are boundary conditions of diffusion. This is rewritten as:

\[\begin{align*} (\frac{1}{\Delta t} \mathbf{I} - \theta \mathbf{D}) \mathbf{u}^{n+1} & = (\frac{1}{\Delta t} \mathbf{I} - (1 - \theta) \mathbf{D}) \mathbf{u}^{n} \\ & - (\alpha_1 \mathbf{c}^n + \alpha_2 \mathbf{c}^{n-1}) \\ & + \theta \mathbf{F}^{n+1} + (1-\theta) \mathbf{F}^{n} \\ & + \theta \mathbf{BC}^{n+1} + (1-\theta) \mathbf{BC}^{n} \\ & - \mathbf{G} \mathbf{p} + \mathbf{y}_p \end{align*}\]

The LU decomposition of the LHS matrix is precomputed in operator_convection_diffusion.jl.

Note that, in constrast to explicit methods, the pressure from previous time steps has an influence on the accuracy of the velocity.

IncompressibleNavierStokes.OneLegMethod — Type

OneLegMethod(; β = 1 // 2, p_add_solve = true)

Explicit one-leg β-method following symmetry-preserving discretization of turbulent flow. See [Verstappen and Veldman (JCP 2003)] for details, or [Direct numerical simulation of turbulence at lower costs (Journal of Engineering Mathematics 1997)].

Formulation:

\[\frac{(\beta + 1/2) u^{n+1} - 2 \beta u^{n} + (\beta - 1/2) u^{n-1}}{\Delta t} = F((1 + \beta) u^n - \beta u^{n-1}).\]

IncompressibleNavierStokes.ExplicitRungeKuttaMethod — Type

ExplicitRungeKuttaMethod(; A, b, c, r, p_add_solve = true)

Explicit Runge Kutta method.

IncompressibleNavierStokes.ImplicitRungeKuttaMethod — Type

ImplicitRungeKuttaMethod(;
    A,
    b,
    c,
    r,
    newton_type = :full,
    maxiter = 10,
    abstol = 1e-14,
    reltol = 1e-14,
    p_add_solve = true,
)

Implicit Runge Kutta method.

The implicit linear system is solved at each time step using Newton's method. The newton_type may be one of the following:

:no: Replace iteration matrix with I/Δt (no Jacobian)

:approximate: Build Jacobian once before iterations only

:full: Build Jacobian at each iteration

IncompressibleNavierStokes.TimeStepper — Type

TimeStepper

Time stepper for solving ODEs.

IncompressibleNavierStokes.AbstractODEMethodCache — Type

AbstractODEMethodCache

ODE method cache.

IncompressibleNavierStokes.ExplicitRungeKuttaCache — Type

ExplicitRungeKuttaCache(; kwargs...)

Explicit Runge-Kutta cache.

IncompressibleNavierStokes.ImplicitRungeKuttaCache — Type

ImplicitRungeKuttaCache(; kwargs...)

Implicit Runge-Kutta cache.

IncompressibleNavierStokes.AdamsBashforthCrankNicolsonCache — Type

AdamsBashforthCrankNicolsonCache(; kwargs...)

Adams-Bashforth Crank-Nicolson cache.

IncompressibleNavierStokes.OneLegCache — Type

OneLegCache(; kwargs...)

One-leg cache.

IncompressibleNavierStokes.change_time_stepper — Function

change_time_stepper(stepper, method) -> TimeStepper

Change ODE method. Return a new stepper.

IncompressibleNavierStokes.isexplicit — Function

isexplicit(method)

Return true if method is explicit, i.e. the value at a certain time step is given explicitly as a function of the previous time steps onl is given explicitly as a function of the previous time steps only.

IncompressibleNavierStokes.lambda_conv_max — Function

lambda_conv_max(method)

Get maximum value of stability region for the convection operator (not a very good indication for the methods that do not include the imaginary axis)

IncompressibleNavierStokes.lambda_diff_max — Function

lambda_diff_max(method)

Get maximum value of stability region for the diffusion operator.

IncompressibleNavierStokes.needs_startup_method — Function

needs_startup_method(method)

Return true if method needs a startup method to generate an initial history.

IncompressibleNavierStokes.nstage — Function

nstage(rk_method)

Get number of stages of the Runge-Kutta method.

IncompressibleNavierStokes.ode_method_cache — Function

ode_method_cache(method, setup)

Get time stepper cache for the given ODE method.

IncompressibleNavierStokes.runge_kutta_method — Function

runge_kutta_method(A, b, c, r; [p_add_solve], [newton_type], [maxiter], [abstol], [reltol])

Get Runge Kutta method. The function checks whether the method is explicit.

p_add_solve: whether to add a pressure solve step to the method.

For implicit RK methods: newton_type, maxiter, abstol, reltol.

IncompressibleNavierStokes.step — Function

step(stepper::AdamsBashforthCrankNicolsonStepper, Δt; bc_vectors = nothing)

Perform one time step with Adams-Bashforth for convection and Crank-Nicolson for diffusion.

Output includes convection terms at tₙ, which will be used in next time step in the Adams-Bashforth part of the method.

Non-mutating/allocating/out-of-place version.

See also step!.

step(stepper::OneLegStepper, Δt; bc_vectors = nothing)

Do one time step using one-leg-β-method.

Non-mutating/allocating/out-of-place version.

See also step!.

step(stepper::ExplicitRungeKuttaStepper, Δt; bc_vectors = nothing)

Perform one time step for the general explicit Runge-Kutta method (ERK).

Dirichlet boundary points are not part of solution vector but are prescribed in a strong manner via the u_bc and v_bc functions.

Non-mutating/allocating/out-of-place version.

See also step!.

step(stepper::ImplicitRungeKuttaStepper, Δt; bc_vectors = nothing)

Do one time step for implicit Runge-Kutta method.

Unsteady Dirichlet boundary points are not part of solution vector but are prescribed in a "strong" manner via the u_bc and v_bc functions.

Non-mutating/allocating/out-of-place version.

See also step!.

IncompressibleNavierStokes.step! — Function

step!(stepper::AdamsBashforthCrankNicolsonStepper, Δt; cache, momentum_cache, bc_vectors = nothing)

Perform one time step with Adams-Bashforth for convection and Crank-Nicolson for diffusion.

Output includes convection terms at tₙ, which will be used in next time step in the Adams-Bashforth part of the method.

Mutating/non-allocating/in-place version.

See also step.

step!(stepper::OneLegStepper, Δt; cache, momentum_cache, bc_vectors = nothing)

Do one time step using one-leg-β-method.

step!(stepper::ExplicitRungeKuttaStepper, Δt; cache, momentum_cache, bc_vectors = nothing)

Perform one time step for the general explicit Runge-Kutta method (ERK).

Dirichlet boundary points are not part of solution vector but are prescribed in a strong manner via the u_bc and v_bc functions.

Mutating/non-allocating/in-place version.

See also step.

step!(stepper::ImplicitRungeKuttaStepper, Δt; cache, momentum_cache, bc_vectors = nothing)

Do one time step for implicit Runge-Kutta method.

Unsteady Dirichlet boundary points are not part of solution vector but are prescribed in a "strong" manner via the u_bc and v_bc functions.

Mutating/non-allocating/in-place version.

See also step.

Utils

IncompressibleNavierStokes.filter_convection — Function

filter_convection(u, diff_matrix, bc, α)

Construct filter for convective terms.

Non-mutating/allocating/out-of-place version.

See also filter_convection!.

IncompressibleNavierStokes.filter_convection! — Function

filter_convection!(ū, u, diff_matrix, bc, α)

Construct filter for convective terms.

Mutating/non-allocating/in-place version.

See also filter_convection.

IncompressibleNavierStokes.get_lims — Function

get_lims(x, n = 1.5)

Get approximate lower and upper limits of a field x based on the mean and standard deviation ($\mu \pm n \sigma$). If x is constant, a margin of 1e-4 is enforced. This is required for contour plotting functions that require a certain range.