Neural closure models

For large eddy simulation (LES), a closure model is required. With IncompressibleNavierStokes, a neural closure model can be trained on filtered DNS data. The discrete DNS equations are given by

\[\begin{split} M u & = 0, \\ \frac{\mathrm{d} u}{\mathrm{d} t} & = F(u) - G p. \end{split}\]

Applying a spatial filter $\Phi$, the extracted large scale components $\bar{u} = \Phi u$ are governed by the equation

\[\begin{split} M \bar{u} & = 0, \\ \frac{\mathrm{d} \bar{u}}{\mathrm{d} t} & = F(\bar{u}) + c - G \bar{p}, \end{split}\]

where the discretizations $M$, $F$, and $G$ are adapted to the size of their inputs and $c = \overline{F(u)} - F(\bar{u})$ is a commutator error. We here assumed that $M$ and $\Phi$ commute, which is the case for face averaging filters. Replacing $c$ with a parameterized closure model $m(\bar{u}, \theta) \approx c$ gives the LES equations for the approximate large scale velocity $\bar{v} \approx \bar{u}$

\[\begin{split} M \bar{v} & = 0, \\ \frac{\mathrm{d} \bar{v}}{\mathrm{d} t} & = F(\bar{v}) + m(\bar{v}, \theta) - G \bar{q}. \end{split}\]

IncompressibleNavierStokes.face_average! — Function

face_average!(v, u, setup, comp)

Average u over volume faces. Put result in v.

IncompressibleNavierStokes.create_les_data — Function

create_les_data(
    T;
    D = 2,
    Re = T(2_000),
    lims = (T(0), T(1)),
    nles = 64,
    compression = 4,
    nsim = 10,
    tburn = T(0.1),
    tsim = T(0.1),
    Δt = T(1e-4),
    ArrayType = Array,
)

Create filtered DNS data.

IncompressibleNavierStokes.create_io_arrays — Function

create_io_arrays(data, setup)

Create $(\bar{u}, c)$ pairs for training.

Training

To improve the model parameters, we exploit exact filtered DNS data $\bar{u}$ and exact commutator errors $c$ obtained through DNS. The model is trained by minimizing the a priori loss function

\[L^\text{prior}(\theta) = \| m(\bar{u}, \theta) - c \|^2,\]

or the a posteriori loss function

\[L^\text{post}(\theta) = \| \bar{v}_\theta - \bar{u} \|^2,\]

where $\bar{v}_\theta$ is the solution to the LES equation for the given parameters $\theta$. The prior loss is easy to evaluate and easy to differentiate, as it does not involve solving the ODE. However, minimizing $L^\text{prior}$ does not take into account the effect of the prediction error on the LES solution error. The posterior loss does, but has a longer computational chain involving solving the LES ODE.

IncompressibleNavierStokes.train — Function

train(
    loss,
    opt,
    θ;
    niter = 100,
    ncallback = 1,
    callback = (i, θ) -> println("Iteration $i of $niter"),
)

Update parameters θ to minimize loss(θ) using the optimiser opt for niter iterations.

Return the a new named tuple (; opt, θ, callbackstate) with updated state and parameters.

IncompressibleNavierStokes.create_randloss — Function

create_randloss(loss, f, x, y; nuse = size(x, 2), device = identity)

Create loss function randloss(θ) that uses a batch of nuse random samples from (x, y) at each evaluation.

The function loss should take inputs like loss(f, x, y, θ).

The batch is moved to device before the loss is evaluated.

IncompressibleNavierStokes.mean_squared_error — Function

mean_squared_error(f, x, y, θ; normalize = y -> sum(abs2, y), λ = sqrt(eps(eltype(x))))

Compute MSE between f(x, θ) and y.

The MSE is further divided by normalize(y).

IncompressibleNavierStokes.relative_error — Function

relative_error(x, y)

Compute average column relative error between matrices x and y.

IncompressibleNavierStokes.create_callback — Function

create_callback(
    f,
    x,
    y;
    state = Point2f[],
    display_each_iteration = false,
)

Create convergence plot for relative error between f(x, θ) and y. At each callback, plot is updated and current error is printed.

If state is nonempty, it also plots previous convergence.

If not using interactive GLMakie window, set display_each_iteration to true.

Neural architectures

We provide two neural architectures: A convolutional neural network (CNN) and a Fourier neural operator (FNO).

IncompressibleNavierStokes.cnn — Function

cnn(setup, r, c, σ, b; kwargs...)

Create CNN closure model. Return a tuple (closure, θ) where θ are the initial parameters and closure(u, θ) predicts the commutator error.

IncompressibleNavierStokes.fno — Function

fno(setup, kmax, c, σ, ψ; rng = Random.default_rng(), kwargs...)

Create FNO closure model. Return a tuple (closure, θ) where θ are the initial parameters and closure(V, θ) predicts the commutator error.

IncompressibleNavierStokes.FourierLayer — Type

FourierLayer(dimension, kmax, cin => cout; σ = identity, init_weight = glorot_uniform)

Fourier layer operating on uniformly discretized functions.

Some important sizes:

dimension: Spatial dimension, e.g. Dimension(2) or Dimension(3).

(nx..., cin, nsample): Input size

(nx..., cout, nsample): Output size

nx = fill(n, dimension()): Number of points in each spatial dimension

n ≥ kmax: Same number of points in each spatial dimension, must be larger than cut-off wavenumber

kmax: Cut-off wavenumber

nsample: Number of input samples (treated independently)