Operators

All discrete operators are built using KernelAbstractions.jl and Cartesian indices, similar to WaterLily.jl. This allows for dimension- and backend-agnostic code. See this blog post for how to write kernels. IncompressibleNavierStokes previously relied on assembling sparse operators to perform the same operations. While being very efficient and also compatible with CUDA (CUSPARSE), storing these matrices in memory is expensive for large 3D problems.

IncompressibleNavierStokes.Offset — Type

δ = Offset{D}()

Carsesian index unit vector in D = 2 or D = 3 dimensions. Calling δ(α) returns a Cartesian index with 1 in the dimension α and zeros elsewhere.

See https://b-fg.github.io/2023/05/07/waterlily-on-gpu.html for writing kernel loops using Cartesian indices.

IncompressibleNavierStokes.divergence! — Function

divergence!(M, u, setup)

Compute divergence of velocity field (in-place version).

IncompressibleNavierStokes.divergence — Function

divergence(u, setup)

Compute divergence of velocity field.

IncompressibleNavierStokes.vorticity — Function

vorticity(u, setup)

Compute vorticity field.

IncompressibleNavierStokes.vorticity! — Function

vorticity!(ω, u, setup)

Compute vorticity field.

IncompressibleNavierStokes.convection! — Function

convection!(F, u, setup)

Compute convective term.

IncompressibleNavierStokes.diffusion! — Function

diffusion!(F, u, setup)

Compute diffusive term.

IncompressibleNavierStokes.bodyforce! — Function

bodyforce!(F, u, setup)

Compute body force.

IncompressibleNavierStokes.momentum! — Function

momentum!(F, u, t, setup)

Right hand side of momentum equations, excluding pressure gradient. Put the result in $F$.

IncompressibleNavierStokes.momentum — Function

momentum(u, t, setup)

Right hand side of momentum equations, excluding pressure gradient.

IncompressibleNavierStokes.laplacian! — Function

laplacian!(L, p, setup)

Compute Laplacian of pressure field (in-place version).

IncompressibleNavierStokes.pressuregradient! — Function

pressuregradient!(G, p, setup)

Compute pressure gradient (in-place).

IncompressibleNavierStokes.pressuregradient — Function

pressuregradient(p, setup)

Compute pressure gradient.

IncompressibleNavierStokes.interpolate_u_p — Function

interpolate_u_p(setup, u)

Interpolate velocity to pressure points.

IncompressibleNavierStokes.interpolate_u_p! — Function

interpolate_u_p!(setup, up, u)

Interpolate velocity to pressure points.

IncompressibleNavierStokes.interpolate_ω_p — Function

interpolate_ω_p(setup, ω)

Interpolate vorticity to pressure points.

IncompressibleNavierStokes.interpolate_ω_p! — Function

interpolate_ω_p!(setup, ωp, ω)

Interpolate vorticity to pressure points.

IncompressibleNavierStokes.Dfield! — Function

Dfield!(d, G, p, setup; ϵ = eps(eltype(p)))

Compute the $D$-field [7] given by

\[D = \frac{2 | \nabla p |}{\nabla^2 p}.\]

IncompressibleNavierStokes.Dfield — Function

Dfield(p, setup)

Compute the $D$-field.

IncompressibleNavierStokes.Qfield! — Function

Qfield!(Q, u, setup; ϵ = eps(eltype(Q)))

Compute $Q$-field [8] given by

\[Q = - \frac{1}{2} \sum_{α, β} \frac{\partial u^α}{\partial x^β} \frac{\partial u^β}{\partial x^α}.\]

IncompressibleNavierStokes.Qfield — Function

Qfield(u, setup)

Compute the $Q$-field.

IncompressibleNavierStokes.kinetic_energy — Function

kinetic_energy(setup, u)

Compute total kinetic energy. The velocity components are interpolated to the volume centers and squared.