Operators

All discrete operators are built using KernelAbstractions.jl and Cartesian indices, similar to WaterLily.jl. This allows for dimension- and backend-agnostic code. See this blog post for how to write kernels. IncompressibleNavierStokes previously relied on assembling sparse operators to perform the same operations. While being very efficient and also compatible with CUDA (CUSPARSE), storing these matrices in memory is expensive for large 3D problems.

IncompressibleNavierStokes.Offset — Type

δ = Offset{D}()

Cartesian index unit vector in D = 2 or D = 3 dimensions. Calling δ(α) returns a Cartesian index with 1 in the dimension α and zeros elsewhere.

See https://b-fg.github.io/2023/05/07/waterlily-on-gpu.html for writing kernel loops using Cartesian indices.

IncompressibleNavierStokes.divergence! — Function

divergence!(div, u, setup)

Compute divergence of velocity field (in-place version).

IncompressibleNavierStokes.divergence — Function

divergence(u, setup)

Compute divergence of velocity field.

IncompressibleNavierStokes.vorticity — Function

vorticity(u, setup)

Compute vorticity field.

IncompressibleNavierStokes.vorticity! — Function

vorticity!(ω, u, setup)

Compute vorticity field.

IncompressibleNavierStokes.convection! — Function

convection!(F, u, setup)

Compute convective term.

IncompressibleNavierStokes.diffusion! — Function

diffusion!(F, u, setup)

Compute diffusive term.

IncompressibleNavierStokes.bodyforce! — Function

bodyforce!(F, u, t, setup)

Compute body force.

IncompressibleNavierStokes.momentum! — Function

momentum!(F, u, t, setup)

Right hand side of momentum equations, excluding pressure gradient. Put the result in $F$.

IncompressibleNavierStokes.momentum — Function

momentum(u, t, setup)

Right hand side of momentum equations, excluding pressure gradient.

IncompressibleNavierStokes.laplacian! — Function

laplacian!(L, p, setup)

Compute Laplacian of pressure field (in-place version).

IncompressibleNavierStokes.laplacian — Function

laplacian(p, setup)

Compute Laplacian of pressure field.

IncompressibleNavierStokes.pressuregradient! — Function

pressuregradient!(G, p, setup)

Compute pressure gradient (in-place).

IncompressibleNavierStokes.pressuregradient — Function

pressuregradient(p, setup)

Compute pressure gradient.

IncompressibleNavierStokes.interpolate_u_p — Function

interpolate_u_p(u, setup)

Interpolate velocity to pressure points.

IncompressibleNavierStokes.interpolate_u_p! — Function

interpolate_u_p!(up, u, setup)

Interpolate velocity to pressure points.

IncompressibleNavierStokes.interpolate_ω_p — Function

interpolate_ω_p(ω, setup)

Interpolate vorticity to pressure points.

IncompressibleNavierStokes.interpolate_ω_p! — Function

interpolate_ω_p!(ωp, ω, setup)

Interpolate vorticity to pressure points.

IncompressibleNavierStokes.Dfield! — Function

Dfield!(d, G, p, setup; ϵ = eps(eltype(p)))

Compute the $D$-field [10] given by

\[D = \frac{2 | \nabla p |}{\nabla^2 p}.\]

IncompressibleNavierStokes.Dfield — Function

Dfield(p, setup)

Compute the $D$-field.

IncompressibleNavierStokes.Qfield! — Function

Qfield!(Q, u, setup)

Compute $Q$-field [11] given by

\[Q = - \frac{1}{2} \sum_{α, β} \frac{\partial u^α}{\partial x^β} \frac{\partial u^β}{\partial x^α}.\]

IncompressibleNavierStokes.Qfield — Function

Qfield(u, setup)

Compute the $Q$-field.

IncompressibleNavierStokes.eig2field! — Function

eig2field!(λ, u, setup; ϵ = eps(eltype(λ)))

Compute the second eigenvalue of $S^2 + \Omega^2$, as proposed by Jeong and Hussain [11].

IncompressibleNavierStokes.eig2field — Function

eig2field(u, setup)

Compute the second eigenvalue of $S^2 + \Omega^2$, as proposed by Jeong and Hussain [11].

IncompressibleNavierStokes.kinetic_energy — Function

kinetic_energy(u, setup; kwargs...)

Compute kinetic energy field $e$ (out-of-place version).

IncompressibleNavierStokes.kinetic_energy! — Function

kinetic_energy!(e, u, setup; interpolate_first = false)

Compute kinetic energy field $e$ (in-place version). If interpolate_first is true, it is given by

\[e_I = \frac{1}{8} \sum_\alpha (u^\alpha_{I + \delta(\alpha) / 2} + u^\alpha_{I - \delta(\alpha) / 2})^2.\]

Otherwise, it is given by

\[e_I = \frac{1}{4} \sum_\alpha (u^\alpha_{I + \delta(\alpha) / 2}^2 + u^\alpha_{I - \delta(\alpha) / 2}^2),\]

as in [12].

IncompressibleNavierStokes.total_kinetic_energy — Function

total_kinetic_energy(setup, u; kwargs...)

Compute total kinetic energy. The velocity components are interpolated to the volume centers and squared.

IncompressibleNavierStokes.tensorbasis — Function

tensorbasis(u, setup)

Compute symmetry tensor basis T[1]-T[11] and invariants V[1]-V[5].

IncompressibleNavierStokes.smagtensor! — Function

smagtensor!(σ, u, θ, setup)

Compute Smagorinsky stress tensors σ[I]. The Smagorinsky constant θ should be a scalar between 0 and 1.

IncompressibleNavierStokes.smagorinsky_closure — Function

m = smagorinsky_closure(setup)

Create Smagorinsky closure model m. The model is called as m(u, θ), where the Smagorinsky constant θ should be a scalar between 0 and 1 (for example θ = 0.1).

IncompressibleNavierStokes.divoftensor! — Function

divoftensor!(s, σ, setup)

Compute divergence of a tensor with all components in the pressure points. The stress tensors should be precomputed and stored in σ.

IncompressibleNavierStokes.tensorbasis! — Function

tensorbasis!(B, V, u, setup)

Compute symmetry tensor basis B[1]-B[11] and invariants V[1]-V[5], as specified in [13] in equations (9) and (11). Note that B[1] corresponds to $T_0$ in the paper, and V to $I$.

IncompressibleNavierStokes.reconstruct! — Function

reconstruct!(u, v, setup_dns, setup_les)

Average fine grid u over coarse volume face. Put result in v.